@@ -9,34 +9,74 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"net/http"
 	"os"
 	"path/filepath"
+	"time"
+
+	"github.com/spf13/cobra"
 
 	"github.com/moov-io/metro2/pkg/file"
 	"github.com/moov-io/metro2/pkg/server"
 	"github.com/moov-io/metro2/pkg/utils"
-	"github.com/spf13/cobra"
 )
 
-var (
-	inputFile = ""
-	rawData   = ""
-)
+// utility functions
+func createReader(cmd *cobra.Command) (io.Reader, error) {
 
-var WebCmd = &cobra.Command{
+	input, err := cmd.Flags().GetString("input")
+	if err != nil {
+		return nil, err
+	}
+
+	if input == "" {
+		path, _ := os.Getwd()
+		input = filepath.Join(path, "metro.json")
+	}
+
+	_, err = os.Stat(input)
+	if os.IsNotExist(err) {
+		return nil, errors.New("invalid input file")
+	}
+	reader, err := os.Open(input)
+	if err != nil {
+		return nil, err
+	}
+
+	return reader, nil
+}
+
+func closeReader(reader io.Reader) {
+	if closer, ok := reader.(io.Closer); ok {
+		closer.Close()
+	}
+}
+
+// commands
+var webCmd = &cobra.Command{
 	Use:   "web",
 	Short: "Launches web server",
 	Long:  "Launches web server",
 	RunE: func(cmd *cobra.Command, args []string) error {
+
 		port, _ := cmd.Flags().GetString("port")
-		fmt.Println("Starting web server on port ", port)
-		listen := "0.0.0.0:" + port
-		h, _ := server.ConfigureHandlers()
+		fmt.Fprintf(os.Stdout, "Starting web server on port %s\n\n", port)
+
+		timeout, _ := time.ParseDuration("30s")
+		handler, _ := server.ConfigureHandlers()
+		serve := &http.Server{
+			Addr:              "0.0.0.0:" + port,
+			Handler:           handler,
+			ReadTimeout:       timeout,
+			ReadHeaderTimeout: timeout,
+			WriteTimeout:      timeout,
+			IdleTimeout:       timeout,
+		}
+
 		test, _ := cmd.Flags().GetBool("test")
 		if !test {
-			err := http.ListenAndServe(listen, h)
-			if err != nil {
+			if err := serve.ListenAndServe(); err != nil {
 				return err
 			}
 		}
@@ -44,24 +84,46 @@ var WebCmd = &cobra.Command{
 	},
 }
 
-var Validate = &cobra.Command{
+var validate = &cobra.Command{
 	Use:   "validator",
 	Short: "Validate metro file",
 	Long:  "Validate an incoming metro file",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		f, err := file.CreateFile([]byte(rawData))
+
+		reader, err := createReader(cmd)
 		if err != nil {
 			return err
 		}
-		return f.Validate()
+		defer closeReader(reader)
+
+		f, err := file.NewFileFromReader(reader)
+		if err != nil {
+			return err
+		}
+
+		err = f.Validate()
+		if err != nil {
+			return err
+		}
+
+		fmt.Fprintf(os.Stdout, "the file is valid \n")
+
+		return nil
 	},
 }
 
-var Print = &cobra.Command{
+var print = &cobra.Command{
 	Use:   "print",
 	Short: "Print metro file",
 	Long:  "Print an incoming metro file with special format (options: metro, json)",
 	RunE: func(cmd *cobra.Command, args []string) error {
+
+		reader, err := createReader(cmd)
+		if err != nil {
+			return err
+		}
+		defer closeReader(reader)
+
 		format, err := cmd.Flags().GetString("format")
 		if err != nil {
 			return err
@@ -75,12 +137,14 @@ var Print = &cobra.Command{
 			}
 		}
 
-		f, err := file.CreateFile([]byte(rawData))
+		f, err := file.NewFileFromReader(reader)
 		if err != nil {
 			return err
 		}
 
+		newline, _ := cmd.Flags().GetBool("newline")
 		output := ""
+
 		if format == utils.MessageJsonFormat {
 			buf, err := json.Marshal(f)
 			if err != nil {
@@ -93,14 +157,16 @@ var Print = &cobra.Command{
 			}
 			output = pretty.String()
 		} else if format == utils.MessageMetroFormat {
-			output = f.String()
+			output = f.String(newline)
 		}
-		fmt.Println(output)
+
+		fmt.Fprintf(os.Stdout, "%s", output)
+
 		return nil
 	},
 }
 
-var Convert = &cobra.Command{
+var convert = &cobra.Command{
 	Use:   "convert [output]",
 	Short: "Convert metro file format",
 	Long:  "Convert an incoming metro file into another format (options: metro, json)",
@@ -111,6 +177,13 @@ var Convert = &cobra.Command{
 		return nil
 	},
 	RunE: func(cmd *cobra.Command, args []string) error {
+
+		reader, err := createReader(cmd)
+		if err != nil {
+			return err
+		}
+		defer closeReader(reader)
+
 		format, err := cmd.Flags().GetString("format")
 		if err != nil {
 			return err
@@ -124,7 +197,7 @@ var Convert = &cobra.Command{
 			}
 		}
 
-		mf, err := file.CreateFile([]byte(rawData))
+		mf, err := file.NewFileFromReader(reader)
 		if err != nil {
 			return err
 		}
@@ -141,6 +214,13 @@ var Convert = &cobra.Command{
 			}
 		}
 
+		newline, _ := cmd.Flags().GetBool("newline")
+		if newType, tErr := cmd.Flags().GetString("type"); tErr == nil {
+			if newType == utils.PackedFileFormat || newType == utils.CharacterFileFormat {
+				mf.SetType(newType)
+			}
+		}
+
 		output := ""
 		if format == utils.MessageJsonFormat {
 			buf, err := json.Marshal(mf)
@@ -154,7 +234,7 @@ var Convert = &cobra.Command{
 			}
 			output = pretty.String()
 		} else if format == utils.MessageMetroFormat {
-			output = mf.String()
+			output = mf.String(newline)
 		}
 
 		f, err := os.Create(args[0])
@@ -167,59 +247,27 @@ var Convert = &cobra.Command{
 	},
 }
 
-var rootCmd = &cobra.Command{
-	Use:   "",
-	Short: "",
-	Long:  "",
-	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-		isWeb := false
-		cmdNames := make([]string, 0)
-		getName := func(c *cobra.Command) {}
-		getName = func(c *cobra.Command) {
-			if c == nil {
-				return
-			}
-			cmdNames = append([]string{c.Name()}, cmdNames...)
-			if c.Name() == "web" {
-				isWeb = true
-			}
-			getName(c.Parent())
-		}
-		getName(cmd)
-
-		if !isWeb {
-			if inputFile == "" {
-				path, _ := os.Getwd()
-				inputFile = filepath.Join(path, "metro.json")
-			}
-			_, err := os.Stat(inputFile)
-			if os.IsNotExist(err) {
-				return errors.New("invalid input file")
-			}
-			f, err := os.Open(inputFile)
-			if err != nil {
-				return err
-			}
-			rawData = utils.ReadFile(f)
-		}
-
-		return nil
-	},
-}
+var rootCmd = &cobra.Command{}
 
 func initRootCmd() {
-	WebCmd.Flags().String("port", "8080", "port of the web server")
-	WebCmd.Flags().BoolP("test", "t", false, "test server")
-	Convert.Flags().String("format", "json", "format of metro file(required)")
-	Convert.Flags().BoolP("generate", "g", false, "generate trailer record")
-	Print.Flags().String("format", "json", "print format")
+	webCmd.Flags().String("port", "8080", "port of the web server")
+	webCmd.Flags().BoolP("test", "t", false, "test server")
+
+	convert.Flags().String("format", "json", "format of metro file(required)")
+	convert.Flags().String("type", "", "file type (character or packed)")
+	convert.Flags().BoolP("generate", "g", false, "generate trailer record")
+	convert.Flags().BoolP("newline", "n", false, "has new line")
+
+	print.Flags().String("format", "json", "print format")
+	print.Flags().BoolP("newline", "n", false, "has new line")
 
 	rootCmd.SilenceUsage = true
-	rootCmd.PersistentFlags().StringVar(&inputFile, "input", "", "input file (default is $PWD/metro.json)")
-	rootCmd.AddCommand(WebCmd)
-	rootCmd.AddCommand(Convert)
-	rootCmd.AddCommand(Print)
-	rootCmd.AddCommand(Validate)
+	rootCmd.PersistentFlags().String("input", "", "input file (default is $PWD/metro.json)")
+
+	rootCmd.AddCommand(webCmd)
+	rootCmd.AddCommand(convert)
+	rootCmd.AddCommand(print)
+	rootCmd.AddCommand(validate)
 }
 
 func main() {
