@@ -1,67 +1,119 @@
 /*
- * METRO2 API
- *
- * Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
+METRO2 API
+
+Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
+
+API version: v1
+*/
+
+// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
 
 package client
 
 import (
-	_context "context"
-	"github.com/antihax/optional"
-	_ioutil "io/ioutil"
-	_nethttp "net/http"
-	_neturl "net/url"
+	"bytes"
+	"context"
+	"io/ioutil"
+	"net/http"
+	"net/url"
 	"os"
 )
 
 // Linger please
 var (
-	_ _context.Context
+	_ context.Context
 )
 
 // Metro2FilesApiService Metro2FilesApi service
 type Metro2FilesApiService service
 
-// ConvertOpts Optional parameters for the method 'Convert'
-type ConvertOpts struct {
-	Format   optional.String
-	Generate optional.Bool
-	File     optional.Interface
+type ApiConvertRequest struct {
+	ctx context.Context
+	ApiService *Metro2FilesApiService
+	format *string
+	format2 *string
+	type_ *string
+	generate *bool
+	newline *bool
+	file **os.File
+}
+
+// print metro2 file type
+func (r ApiConvertRequest) Format(format string) ApiConvertRequest {
+	r.format = &format
+	return r
+}
+// format of metro file
+func (r ApiConvertRequest) Format2(format2 string) ApiConvertRequest {
+	r.format2 = &format2
+	return r
+}
+// metro file type
+func (r ApiConvertRequest) Type_(type_ string) ApiConvertRequest {
+	r.type_ = &type_
+	return r
+}
+// generate new trailer record
+func (r ApiConvertRequest) Generate(generate bool) ApiConvertRequest {
+	r.generate = &generate
+	return r
+}
+// has new line
+func (r ApiConvertRequest) Newline(newline bool) ApiConvertRequest {
+	r.newline = &newline
+	return r
+}
+// metro2 file to upload
+func (r ApiConvertRequest) File(file *os.File) ApiConvertRequest {
+	r.file = &file
+	return r
+}
+
+func (r ApiConvertRequest) Execute() (**os.File, *http.Response, error) {
+	return r.ApiService.ConvertExecute(r)
 }
 
 /*
 Convert Convert metro2 file
+
 Convert from original metro2 file to new metro2 file
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param optional nil or *ConvertOpts - Optional Parameters:
- * @param "Format" (optional.String) -  print metro2 file type
- * @param "Generate" (optional.Bool) -  generate new trailer record
- * @param "File" (optional.Interface of *os.File) -  metro2 file to upload
-@return *os.File
+
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @return ApiConvertRequest
 */
-func (a *Metro2FilesApiService) Convert(ctx _context.Context, localVarOptionals *ConvertOpts) (*os.File, *_nethttp.Response, error) {
+func (a *Metro2FilesApiService) Convert(ctx context.Context) ApiConvertRequest {
+	return ApiConvertRequest{
+		ApiService: a,
+		ctx: ctx,
+	}
+}
+
+// Execute executes the request
+//  @return *os.File
+func (a *Metro2FilesApiService) ConvertExecute(r ApiConvertRequest) (**os.File, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  *os.File
+		formFiles            []formFile
+		localVarReturnValue  **os.File
 	)
 
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/convert"
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "Metro2FilesApiService.Convert")
+	if err != nil {
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
+	}
+
+	localVarPath := localBasePath + "/convert"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
+	if r.format != nil {
+		localVarQueryParams.Add("format", parameterToString(*r.format, ""))
+	}
 	// to determine the Content-Type header
-	localVarHTTPContentTypes := []string{"multipart/form-data"}
+	localVarHTTPContentTypes := []string{"multipart/form-data", "application/json"}
 
 	// set Content-Type header
 	localVarHTTPContentType := selectHeaderContentType(localVarHTTPContentTypes)
@@ -77,57 +129,56 @@ func (a *Metro2FilesApiService) Convert(ctx _context.Context, localVarOptionals 
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-	if localVarOptionals != nil && localVarOptionals.Format.IsSet() {
-		localVarFormParams.Add("format", parameterToString(localVarOptionals.Format.Value(), ""))
+	if r.format2 != nil {
+		localVarFormParams.Add("format", parameterToString(*r.format2, ""))
 	}
-	if localVarOptionals != nil && localVarOptionals.Generate.IsSet() {
-		localVarFormParams.Add("generate", parameterToString(localVarOptionals.Generate.Value(), ""))
+	if r.type_ != nil {
+		localVarFormParams.Add("type", parameterToString(*r.type_, ""))
 	}
-	localVarFormFileName = "file"
-	var localVarFile *os.File
-	if localVarOptionals != nil && localVarOptionals.File.IsSet() {
-		localVarFileOk := false
-		localVarFile, localVarFileOk = localVarOptionals.File.Value().(*os.File)
-		if !localVarFileOk {
-			return localVarReturnValue, nil, reportError("file should be *os.File")
-		}
+	if r.generate != nil {
+		localVarFormParams.Add("generate", parameterToString(*r.generate, ""))
 	}
-	if localVarFile != nil {
-		fbs, _ := _ioutil.ReadAll(localVarFile)
-		localVarFileBytes = fbs
-		localVarFileName = localVarFile.Name()
-		localVarFile.Close()
+	if r.newline != nil {
+		localVarFormParams.Add("newline", parameterToString(*r.newline, ""))
 	}
-	r, err := a.client.prepareRequest(ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
+	var fileLocalVarFormFileName string
+	var fileLocalVarFileName     string
+	var fileLocalVarFileBytes    []byte
+
+	fileLocalVarFormFileName = "file"
+
+	var fileLocalVarFile *os.File
+	if r.file != nil {
+		fileLocalVarFile = *r.file
+	}
+	if fileLocalVarFile != nil {
+		fbs, _ := ioutil.ReadAll(fileLocalVarFile)
+		fileLocalVarFileBytes = fbs
+		fileLocalVarFileName = fileLocalVarFile.Name()
+		fileLocalVarFile.Close()
+	}
+	formFiles = append(formFiles, formFile{fileBytes: fileLocalVarFileBytes, fileName: fileLocalVarFileName, formFileName: fileLocalVarFormFileName})
+	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
 	}
 
-	localVarHTTPResponse, err := a.client.callAPI(r)
+	localVarHTTPResponse, err := a.client.callAPI(req)
 	if err != nil || localVarHTTPResponse == nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
+	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
-		}
-		if localVarHTTPResponse.StatusCode == 200 {
-			var v *os.File
-			err = a.client.decode(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-			if err != nil {
-				newErr.error = err.Error()
-				return localVarReturnValue, localVarHTTPResponse, newErr
-			}
-			newErr.model = v
-			return localVarReturnValue, localVarHTTPResponse, newErr
 		}
 		if localVarHTTPResponse.StatusCode == 400 {
 			var v string
@@ -153,7 +204,7 @@ func (a *Metro2FilesApiService) Convert(ctx _context.Context, localVarOptionals 
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -163,28 +214,51 @@ func (a *Metro2FilesApiService) Convert(ctx _context.Context, localVarOptionals 
 	return localVarReturnValue, localVarHTTPResponse, nil
 }
 
+type ApiHealthRequest struct {
+	ctx context.Context
+	ApiService *Metro2FilesApiService
+}
+
+
+func (r ApiHealthRequest) Execute() (string, *http.Response, error) {
+	return r.ApiService.HealthExecute(r)
+}
+
 /*
 Health health metro2 service
+
 Check the metro2 service to check if running
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
-@return string
+
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @return ApiHealthRequest
 */
-func (a *Metro2FilesApiService) Health(ctx _context.Context) (string, *_nethttp.Response, error) {
+func (a *Metro2FilesApiService) Health(ctx context.Context) ApiHealthRequest {
+	return ApiHealthRequest{
+		ApiService: a,
+		ctx: ctx,
+	}
+}
+
+// Execute executes the request
+//  @return string
+func (a *Metro2FilesApiService) HealthExecute(r ApiHealthRequest) (string, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
+		formFiles            []formFile
 		localVarReturnValue  string
 	)
 
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/health"
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "Metro2FilesApiService.Health")
+	if err != nil {
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
+	}
+
+	localVarPath := localBasePath + "/health"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
@@ -203,42 +277,34 @@ func (a *Metro2FilesApiService) Health(ctx _context.Context) (string, *_nethttp.
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-	r, err := a.client.prepareRequest(ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
+	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
 	}
 
-	localVarHTTPResponse, err := a.client.callAPI(r)
+	localVarHTTPResponse, err := a.client.callAPI(req)
 	if err != nil || localVarHTTPResponse == nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
+	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
-		}
-		if localVarHTTPResponse.StatusCode == 200 {
-			var v string
-			err = a.client.decode(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-			if err != nil {
-				newErr.error = err.Error()
-				return localVarReturnValue, localVarHTTPResponse, newErr
-			}
-			newErr.model = v
 		}
 		return localVarReturnValue, localVarHTTPResponse, newErr
 	}
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -248,40 +314,75 @@ func (a *Metro2FilesApiService) Health(ctx _context.Context) (string, *_nethttp.
 	return localVarReturnValue, localVarHTTPResponse, nil
 }
 
-// PrintOpts Optional parameters for the method 'Print'
-type PrintOpts struct {
-	Format optional.String
-	File   optional.Interface
+type ApiPrintRequest struct {
+	ctx context.Context
+	ApiService *Metro2FilesApiService
+	format *string
+	format2 *string
+	file **os.File
+}
+
+// print metro2 file type
+func (r ApiPrintRequest) Format(format string) ApiPrintRequest {
+	r.format = &format
+	return r
+}
+// print metro2 file type
+func (r ApiPrintRequest) Format2(format2 string) ApiPrintRequest {
+	r.format2 = &format2
+	return r
+}
+// metro2 file to upload
+func (r ApiPrintRequest) File(file *os.File) ApiPrintRequest {
+	r.file = &file
+	return r
+}
+
+func (r ApiPrintRequest) Execute() (string, *http.Response, error) {
+	return r.ApiService.PrintExecute(r)
 }
 
 /*
 Print Print metro2 file with specific format
+
 Print metro2 file with requested file format.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param optional nil or *PrintOpts - Optional Parameters:
- * @param "Format" (optional.String) -  print metro2 file type
- * @param "File" (optional.Interface of *os.File) -  metro2 file to upload
-@return string
+
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @return ApiPrintRequest
 */
-func (a *Metro2FilesApiService) Print(ctx _context.Context, localVarOptionals *PrintOpts) (string, *_nethttp.Response, error) {
+func (a *Metro2FilesApiService) Print(ctx context.Context) ApiPrintRequest {
+	return ApiPrintRequest{
+		ApiService: a,
+		ctx: ctx,
+	}
+}
+
+// Execute executes the request
+//  @return string
+func (a *Metro2FilesApiService) PrintExecute(r ApiPrintRequest) (string, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
+		formFiles            []formFile
 		localVarReturnValue  string
 	)
 
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/print"
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "Metro2FilesApiService.Print")
+	if err != nil {
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
+	}
+
+	localVarPath := localBasePath + "/print"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
+	if r.format != nil {
+		localVarQueryParams.Add("format", parameterToString(*r.format, ""))
+	}
 	// to determine the Content-Type header
-	localVarHTTPContentTypes := []string{"multipart/form-data"}
+	localVarHTTPContentTypes := []string{"multipart/form-data", "application/json"}
 
 	// set Content-Type header
 	localVarHTTPContentType := selectHeaderContentType(localVarHTTPContentTypes)
@@ -297,54 +398,47 @@ func (a *Metro2FilesApiService) Print(ctx _context.Context, localVarOptionals *P
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-	if localVarOptionals != nil && localVarOptionals.Format.IsSet() {
-		localVarFormParams.Add("format", parameterToString(localVarOptionals.Format.Value(), ""))
+	if r.format2 != nil {
+		localVarFormParams.Add("format", parameterToString(*r.format2, ""))
 	}
-	localVarFormFileName = "file"
-	var localVarFile *os.File
-	if localVarOptionals != nil && localVarOptionals.File.IsSet() {
-		localVarFileOk := false
-		localVarFile, localVarFileOk = localVarOptionals.File.Value().(*os.File)
-		if !localVarFileOk {
-			return localVarReturnValue, nil, reportError("file should be *os.File")
-		}
+	var fileLocalVarFormFileName string
+	var fileLocalVarFileName     string
+	var fileLocalVarFileBytes    []byte
+
+	fileLocalVarFormFileName = "file"
+
+	var fileLocalVarFile *os.File
+	if r.file != nil {
+		fileLocalVarFile = *r.file
 	}
-	if localVarFile != nil {
-		fbs, _ := _ioutil.ReadAll(localVarFile)
-		localVarFileBytes = fbs
-		localVarFileName = localVarFile.Name()
-		localVarFile.Close()
+	if fileLocalVarFile != nil {
+		fbs, _ := ioutil.ReadAll(fileLocalVarFile)
+		fileLocalVarFileBytes = fbs
+		fileLocalVarFileName = fileLocalVarFile.Name()
+		fileLocalVarFile.Close()
 	}
-	r, err := a.client.prepareRequest(ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
+	formFiles = append(formFiles, formFile{fileBytes: fileLocalVarFileBytes, fileName: fileLocalVarFileName, formFileName: fileLocalVarFormFileName})
+	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
 	}
 
-	localVarHTTPResponse, err := a.client.callAPI(r)
+	localVarHTTPResponse, err := a.client.callAPI(req)
 	if err != nil || localVarHTTPResponse == nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
+	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
-		}
-		if localVarHTTPResponse.StatusCode == 200 {
-			var v string
-			err = a.client.decode(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-			if err != nil {
-				newErr.error = err.Error()
-				return localVarReturnValue, localVarHTTPResponse, newErr
-			}
-			newErr.model = v
-			return localVarReturnValue, localVarHTTPResponse, newErr
 		}
 		if localVarHTTPResponse.StatusCode == 400 {
 			var v string
@@ -360,7 +454,7 @@ func (a *Metro2FilesApiService) Print(ctx _context.Context, localVarOptionals *P
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -370,35 +464,57 @@ func (a *Metro2FilesApiService) Print(ctx _context.Context, localVarOptionals *P
 	return localVarReturnValue, localVarHTTPResponse, nil
 }
 
-// ValidatorOpts Optional parameters for the method 'Validator'
-type ValidatorOpts struct {
-	File optional.Interface
+type ApiValidatorRequest struct {
+	ctx context.Context
+	ApiService *Metro2FilesApiService
+	file **os.File
+}
+
+// metro2 file to upload
+func (r ApiValidatorRequest) File(file *os.File) ApiValidatorRequest {
+	r.file = &file
+	return r
+}
+
+func (r ApiValidatorRequest) Execute() (string, *http.Response, error) {
+	return r.ApiService.ValidatorExecute(r)
 }
 
 /*
 Validator Validate metro2 file
+
 Validation of metro2 file.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param optional nil or *ValidatorOpts - Optional Parameters:
- * @param "File" (optional.Interface of *os.File) -  metro2 file to upload
-@return string
+
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @return ApiValidatorRequest
 */
-func (a *Metro2FilesApiService) Validator(ctx _context.Context, localVarOptionals *ValidatorOpts) (string, *_nethttp.Response, error) {
+func (a *Metro2FilesApiService) Validator(ctx context.Context) ApiValidatorRequest {
+	return ApiValidatorRequest{
+		ApiService: a,
+		ctx: ctx,
+	}
+}
+
+// Execute executes the request
+//  @return string
+func (a *Metro2FilesApiService) ValidatorExecute(r ApiValidatorRequest) (string, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
+		formFiles            []formFile
 		localVarReturnValue  string
 	)
 
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath + "/validator"
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "Metro2FilesApiService.Validator")
+	if err != nil {
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
+	}
+
+	localVarPath := localBasePath + "/validator"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{"multipart/form-data"}
@@ -417,51 +533,44 @@ func (a *Metro2FilesApiService) Validator(ctx _context.Context, localVarOptional
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-	localVarFormFileName = "file"
-	var localVarFile *os.File
-	if localVarOptionals != nil && localVarOptionals.File.IsSet() {
-		localVarFileOk := false
-		localVarFile, localVarFileOk = localVarOptionals.File.Value().(*os.File)
-		if !localVarFileOk {
-			return localVarReturnValue, nil, reportError("file should be *os.File")
-		}
+	var fileLocalVarFormFileName string
+	var fileLocalVarFileName     string
+	var fileLocalVarFileBytes    []byte
+
+	fileLocalVarFormFileName = "file"
+
+	var fileLocalVarFile *os.File
+	if r.file != nil {
+		fileLocalVarFile = *r.file
 	}
-	if localVarFile != nil {
-		fbs, _ := _ioutil.ReadAll(localVarFile)
-		localVarFileBytes = fbs
-		localVarFileName = localVarFile.Name()
-		localVarFile.Close()
+	if fileLocalVarFile != nil {
+		fbs, _ := ioutil.ReadAll(fileLocalVarFile)
+		fileLocalVarFileBytes = fbs
+		fileLocalVarFileName = fileLocalVarFile.Name()
+		fileLocalVarFile.Close()
 	}
-	r, err := a.client.prepareRequest(ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
+	formFiles = append(formFiles, formFile{fileBytes: fileLocalVarFileBytes, fileName: fileLocalVarFileName, formFileName: fileLocalVarFormFileName})
+	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
 	}
 
-	localVarHTTPResponse, err := a.client.callAPI(r)
+	localVarHTTPResponse, err := a.client.callAPI(req)
 	if err != nil || localVarHTTPResponse == nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
+	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
-		}
-		if localVarHTTPResponse.StatusCode == 200 {
-			var v string
-			err = a.client.decode(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-			if err != nil {
-				newErr.error = err.Error()
-				return localVarReturnValue, localVarHTTPResponse, newErr
-			}
-			newErr.model = v
-			return localVarReturnValue, localVarHTTPResponse, newErr
 		}
 		if localVarHTTPResponse.StatusCode == 400 {
 			var v string
@@ -487,7 +596,7 @@ func (a *Metro2FilesApiService) Validator(ctx _context.Context, localVarOptional
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
