@@ -1,28 +1,398 @@
 /*
- * METRO2 API
- *
- * Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
+METRO2 API
+
+Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
+
+API version: v1
+*/
+
+// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
 
 package client
 
 import (
+	"encoding/json"
 	"time"
 )
 
 // J1Segment struct for J1Segment
 type J1Segment struct {
-	SegmentIdentifier            string    `json:"segmentIdentifier"`
-	Surname                      string    `json:"surname"`
-	FirstName                    string    `json:"firstName"`
-	MiddleName                   string    `json:"middleName,omitempty"`
-	GenerationCode               string    `json:"generationCode,omitempty"`
-	SocialSecurityNumber         int32     `json:"socialSecurityNumber"`
-	DateBirth                    time.Time `json:"dateBirth"`
-	TelephoneNumber              int64     `json:"telephoneNumber,omitempty"`
-	EcoaCode                     string    `json:"ecoaCode"`
-	ConsumerInformationIndicator string    `json:"consumerInformationIndicator,omitempty"`
+	SegmentIdentifier string `json:"segmentIdentifier"`
+	Surname string `json:"surname"`
+	FirstName string `json:"firstName"`
+	MiddleName *string `json:"middleName,omitempty"`
+	GenerationCode *string `json:"generationCode,omitempty"`
+	SocialSecurityNumber int32 `json:"socialSecurityNumber"`
+	DateBirth time.Time `json:"dateBirth"`
+	TelephoneNumber *int64 `json:"telephoneNumber,omitempty"`
+	EcoaCode string `json:"ecoaCode"`
+	ConsumerInformationIndicator *string `json:"consumerInformationIndicator,omitempty"`
 }
+
+// NewJ1Segment instantiates a new J1Segment object
+// This constructor will assign default values to properties that have it defined,
+// and makes sure properties required by API are set, but the set of arguments
+// will change when the set of required properties is changed
+func NewJ1Segment(segmentIdentifier string, surname string, firstName string, socialSecurityNumber int32, dateBirth time.Time, ecoaCode string) *J1Segment {
+	this := J1Segment{}
+	this.SegmentIdentifier = segmentIdentifier
+	this.Surname = surname
+	this.FirstName = firstName
+	this.SocialSecurityNumber = socialSecurityNumber
+	this.DateBirth = dateBirth
+	this.EcoaCode = ecoaCode
+	return &this
+}
+
+// NewJ1SegmentWithDefaults instantiates a new J1Segment object
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewJ1SegmentWithDefaults() *J1Segment {
+	this := J1Segment{}
+	return &this
+}
+
+// GetSegmentIdentifier returns the SegmentIdentifier field value
+func (o *J1Segment) GetSegmentIdentifier() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.SegmentIdentifier
+}
+
+// GetSegmentIdentifierOk returns a tuple with the SegmentIdentifier field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetSegmentIdentifierOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.SegmentIdentifier, true
+}
+
+// SetSegmentIdentifier sets field value
+func (o *J1Segment) SetSegmentIdentifier(v string) {
+	o.SegmentIdentifier = v
+}
+
+// GetSurname returns the Surname field value
+func (o *J1Segment) GetSurname() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.Surname
+}
+
+// GetSurnameOk returns a tuple with the Surname field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetSurnameOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.Surname, true
+}
+
+// SetSurname sets field value
+func (o *J1Segment) SetSurname(v string) {
+	o.Surname = v
+}
+
+// GetFirstName returns the FirstName field value
+func (o *J1Segment) GetFirstName() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.FirstName
+}
+
+// GetFirstNameOk returns a tuple with the FirstName field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetFirstNameOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.FirstName, true
+}
+
+// SetFirstName sets field value
+func (o *J1Segment) SetFirstName(v string) {
+	o.FirstName = v
+}
+
+// GetMiddleName returns the MiddleName field value if set, zero value otherwise.
+func (o *J1Segment) GetMiddleName() string {
+	if o == nil || o.MiddleName == nil {
+		var ret string
+		return ret
+	}
+	return *o.MiddleName
+}
+
+// GetMiddleNameOk returns a tuple with the MiddleName field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetMiddleNameOk() (*string, bool) {
+	if o == nil || o.MiddleName == nil {
+		return nil, false
+	}
+	return o.MiddleName, true
+}
+
+// HasMiddleName returns a boolean if a field has been set.
+func (o *J1Segment) HasMiddleName() bool {
+	if o != nil && o.MiddleName != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetMiddleName gets a reference to the given string and assigns it to the MiddleName field.
+func (o *J1Segment) SetMiddleName(v string) {
+	o.MiddleName = &v
+}
+
+// GetGenerationCode returns the GenerationCode field value if set, zero value otherwise.
+func (o *J1Segment) GetGenerationCode() string {
+	if o == nil || o.GenerationCode == nil {
+		var ret string
+		return ret
+	}
+	return *o.GenerationCode
+}
+
+// GetGenerationCodeOk returns a tuple with the GenerationCode field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetGenerationCodeOk() (*string, bool) {
+	if o == nil || o.GenerationCode == nil {
+		return nil, false
+	}
+	return o.GenerationCode, true
+}
+
+// HasGenerationCode returns a boolean if a field has been set.
+func (o *J1Segment) HasGenerationCode() bool {
+	if o != nil && o.GenerationCode != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetGenerationCode gets a reference to the given string and assigns it to the GenerationCode field.
+func (o *J1Segment) SetGenerationCode(v string) {
+	o.GenerationCode = &v
+}
+
+// GetSocialSecurityNumber returns the SocialSecurityNumber field value
+func (o *J1Segment) GetSocialSecurityNumber() int32 {
+	if o == nil {
+		var ret int32
+		return ret
+	}
+
+	return o.SocialSecurityNumber
+}
+
+// GetSocialSecurityNumberOk returns a tuple with the SocialSecurityNumber field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetSocialSecurityNumberOk() (*int32, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.SocialSecurityNumber, true
+}
+
+// SetSocialSecurityNumber sets field value
+func (o *J1Segment) SetSocialSecurityNumber(v int32) {
+	o.SocialSecurityNumber = v
+}
+
+// GetDateBirth returns the DateBirth field value
+func (o *J1Segment) GetDateBirth() time.Time {
+	if o == nil {
+		var ret time.Time
+		return ret
+	}
+
+	return o.DateBirth
+}
+
+// GetDateBirthOk returns a tuple with the DateBirth field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetDateBirthOk() (*time.Time, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.DateBirth, true
+}
+
+// SetDateBirth sets field value
+func (o *J1Segment) SetDateBirth(v time.Time) {
+	o.DateBirth = v
+}
+
+// GetTelephoneNumber returns the TelephoneNumber field value if set, zero value otherwise.
+func (o *J1Segment) GetTelephoneNumber() int64 {
+	if o == nil || o.TelephoneNumber == nil {
+		var ret int64
+		return ret
+	}
+	return *o.TelephoneNumber
+}
+
+// GetTelephoneNumberOk returns a tuple with the TelephoneNumber field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetTelephoneNumberOk() (*int64, bool) {
+	if o == nil || o.TelephoneNumber == nil {
+		return nil, false
+	}
+	return o.TelephoneNumber, true
+}
+
+// HasTelephoneNumber returns a boolean if a field has been set.
+func (o *J1Segment) HasTelephoneNumber() bool {
+	if o != nil && o.TelephoneNumber != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTelephoneNumber gets a reference to the given int64 and assigns it to the TelephoneNumber field.
+func (o *J1Segment) SetTelephoneNumber(v int64) {
+	o.TelephoneNumber = &v
+}
+
+// GetEcoaCode returns the EcoaCode field value
+func (o *J1Segment) GetEcoaCode() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.EcoaCode
+}
+
+// GetEcoaCodeOk returns a tuple with the EcoaCode field value
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetEcoaCodeOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.EcoaCode, true
+}
+
+// SetEcoaCode sets field value
+func (o *J1Segment) SetEcoaCode(v string) {
+	o.EcoaCode = v
+}
+
+// GetConsumerInformationIndicator returns the ConsumerInformationIndicator field value if set, zero value otherwise.
+func (o *J1Segment) GetConsumerInformationIndicator() string {
+	if o == nil || o.ConsumerInformationIndicator == nil {
+		var ret string
+		return ret
+	}
+	return *o.ConsumerInformationIndicator
+}
+
+// GetConsumerInformationIndicatorOk returns a tuple with the ConsumerInformationIndicator field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *J1Segment) GetConsumerInformationIndicatorOk() (*string, bool) {
+	if o == nil || o.ConsumerInformationIndicator == nil {
+		return nil, false
+	}
+	return o.ConsumerInformationIndicator, true
+}
+
+// HasConsumerInformationIndicator returns a boolean if a field has been set.
+func (o *J1Segment) HasConsumerInformationIndicator() bool {
+	if o != nil && o.ConsumerInformationIndicator != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetConsumerInformationIndicator gets a reference to the given string and assigns it to the ConsumerInformationIndicator field.
+func (o *J1Segment) SetConsumerInformationIndicator(v string) {
+	o.ConsumerInformationIndicator = &v
+}
+
+func (o J1Segment) MarshalJSON() ([]byte, error) {
+	toSerialize := map[string]interface{}{}
+	if true {
+		toSerialize["segmentIdentifier"] = o.SegmentIdentifier
+	}
+	if true {
+		toSerialize["surname"] = o.Surname
+	}
+	if true {
+		toSerialize["firstName"] = o.FirstName
+	}
+	if o.MiddleName != nil {
+		toSerialize["middleName"] = o.MiddleName
+	}
+	if o.GenerationCode != nil {
+		toSerialize["generationCode"] = o.GenerationCode
+	}
+	if true {
+		toSerialize["socialSecurityNumber"] = o.SocialSecurityNumber
+	}
+	if true {
+		toSerialize["dateBirth"] = o.DateBirth
+	}
+	if o.TelephoneNumber != nil {
+		toSerialize["telephoneNumber"] = o.TelephoneNumber
+	}
+	if true {
+		toSerialize["ecoaCode"] = o.EcoaCode
+	}
+	if o.ConsumerInformationIndicator != nil {
+		toSerialize["consumerInformationIndicator"] = o.ConsumerInformationIndicator
+	}
+	return json.Marshal(toSerialize)
+}
+
+type NullableJ1Segment struct {
+	value *J1Segment
+	isSet bool
+}
+
+func (v NullableJ1Segment) Get() *J1Segment {
+	return v.value
+}
+
+func (v *NullableJ1Segment) Set(val *J1Segment) {
+	v.value = val
+	v.isSet = true
+}
+
+func (v NullableJ1Segment) IsSet() bool {
+	return v.isSet
+}
+
+func (v *NullableJ1Segment) Unset() {
+	v.value = nil
+	v.isSet = false
+}
+
+func NewNullableJ1Segment(val *J1Segment) *NullableJ1Segment {
+	return &NullableJ1Segment{value: val, isSet: true}
+}
+
+func (v NullableJ1Segment) MarshalJSON() ([]byte, error) {
+	return json.Marshal(v.value)
+}
+
+func (v *NullableJ1Segment) UnmarshalJSON(src []byte) error {
+	v.isSet = true
+	return json.Unmarshal(src, &v.value)
+}
+
+
