@@ -1,18 +1,216 @@
 /*
- * METRO2 API
- *
- * Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
+METRO2 API
+
+Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
+
+API version: v1
+*/
+
+// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
 
 package client
 
+import (
+	"encoding/json"
+)
+
 // K3Segment struct for K3Segment
 type K3Segment struct {
-	SegmentIdentifier            string `json:"segmentIdentifier"`
-	AgencyIdentifier             int32  `json:"agencyIdentifier,omitempty"`
-	AccountNumber                string `json:"accountNumber,omitempty"`
-	MortgageIdentificationNumber string `json:"mortgageIdentificationNumber,omitempty"`
+	SegmentIdentifier string `json:"segmentIdentifier"`
+	AgencyIdentifier *int32 `json:"agencyIdentifier,omitempty"`
+	AccountNumber *string `json:"accountNumber,omitempty"`
+	MortgageIdentificationNumber *string `json:"mortgageIdentificationNumber,omitempty"`
 }
+
+// NewK3Segment instantiates a new K3Segment object
+// This constructor will assign default values to properties that have it defined,
+// and makes sure properties required by API are set, but the set of arguments
+// will change when the set of required properties is changed
+func NewK3Segment(segmentIdentifier string) *K3Segment {
+	this := K3Segment{}
+	this.SegmentIdentifier = segmentIdentifier
+	return &this
+}
+
+// NewK3SegmentWithDefaults instantiates a new K3Segment object
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewK3SegmentWithDefaults() *K3Segment {
+	this := K3Segment{}
+	return &this
+}
+
+// GetSegmentIdentifier returns the SegmentIdentifier field value
+func (o *K3Segment) GetSegmentIdentifier() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.SegmentIdentifier
+}
+
+// GetSegmentIdentifierOk returns a tuple with the SegmentIdentifier field value
+// and a boolean to check if the value has been set.
+func (o *K3Segment) GetSegmentIdentifierOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.SegmentIdentifier, true
+}
+
+// SetSegmentIdentifier sets field value
+func (o *K3Segment) SetSegmentIdentifier(v string) {
+	o.SegmentIdentifier = v
+}
+
+// GetAgencyIdentifier returns the AgencyIdentifier field value if set, zero value otherwise.
+func (o *K3Segment) GetAgencyIdentifier() int32 {
+	if o == nil || o.AgencyIdentifier == nil {
+		var ret int32
+		return ret
+	}
+	return *o.AgencyIdentifier
+}
+
+// GetAgencyIdentifierOk returns a tuple with the AgencyIdentifier field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *K3Segment) GetAgencyIdentifierOk() (*int32, bool) {
+	if o == nil || o.AgencyIdentifier == nil {
+		return nil, false
+	}
+	return o.AgencyIdentifier, true
+}
+
+// HasAgencyIdentifier returns a boolean if a field has been set.
+func (o *K3Segment) HasAgencyIdentifier() bool {
+	if o != nil && o.AgencyIdentifier != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetAgencyIdentifier gets a reference to the given int32 and assigns it to the AgencyIdentifier field.
+func (o *K3Segment) SetAgencyIdentifier(v int32) {
+	o.AgencyIdentifier = &v
+}
+
+// GetAccountNumber returns the AccountNumber field value if set, zero value otherwise.
+func (o *K3Segment) GetAccountNumber() string {
+	if o == nil || o.AccountNumber == nil {
+		var ret string
+		return ret
+	}
+	return *o.AccountNumber
+}
+
+// GetAccountNumberOk returns a tuple with the AccountNumber field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *K3Segment) GetAccountNumberOk() (*string, bool) {
+	if o == nil || o.AccountNumber == nil {
+		return nil, false
+	}
+	return o.AccountNumber, true
+}
+
+// HasAccountNumber returns a boolean if a field has been set.
+func (o *K3Segment) HasAccountNumber() bool {
+	if o != nil && o.AccountNumber != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetAccountNumber gets a reference to the given string and assigns it to the AccountNumber field.
+func (o *K3Segment) SetAccountNumber(v string) {
+	o.AccountNumber = &v
+}
+
+// GetMortgageIdentificationNumber returns the MortgageIdentificationNumber field value if set, zero value otherwise.
+func (o *K3Segment) GetMortgageIdentificationNumber() string {
+	if o == nil || o.MortgageIdentificationNumber == nil {
+		var ret string
+		return ret
+	}
+	return *o.MortgageIdentificationNumber
+}
+
+// GetMortgageIdentificationNumberOk returns a tuple with the MortgageIdentificationNumber field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *K3Segment) GetMortgageIdentificationNumberOk() (*string, bool) {
+	if o == nil || o.MortgageIdentificationNumber == nil {
+		return nil, false
+	}
+	return o.MortgageIdentificationNumber, true
+}
+
+// HasMortgageIdentificationNumber returns a boolean if a field has been set.
+func (o *K3Segment) HasMortgageIdentificationNumber() bool {
+	if o != nil && o.MortgageIdentificationNumber != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetMortgageIdentificationNumber gets a reference to the given string and assigns it to the MortgageIdentificationNumber field.
+func (o *K3Segment) SetMortgageIdentificationNumber(v string) {
+	o.MortgageIdentificationNumber = &v
+}
+
+func (o K3Segment) MarshalJSON() ([]byte, error) {
+	toSerialize := map[string]interface{}{}
+	if true {
+		toSerialize["segmentIdentifier"] = o.SegmentIdentifier
+	}
+	if o.AgencyIdentifier != nil {
+		toSerialize["agencyIdentifier"] = o.AgencyIdentifier
+	}
+	if o.AccountNumber != nil {
+		toSerialize["accountNumber"] = o.AccountNumber
+	}
+	if o.MortgageIdentificationNumber != nil {
+		toSerialize["mortgageIdentificationNumber"] = o.MortgageIdentificationNumber
+	}
+	return json.Marshal(toSerialize)
+}
+
+type NullableK3Segment struct {
+	value *K3Segment
+	isSet bool
+}
+
+func (v NullableK3Segment) Get() *K3Segment {
+	return v.value
+}
+
+func (v *NullableK3Segment) Set(val *K3Segment) {
+	v.value = val
+	v.isSet = true
+}
+
+func (v NullableK3Segment) IsSet() bool {
+	return v.isSet
+}
+
+func (v *NullableK3Segment) Unset() {
+	v.value = nil
+	v.isSet = false
+}
+
+func NewNullableK3Segment(val *K3Segment) *NullableK3Segment {
+	return &NullableK3Segment{value: val, isSet: true}
+}
+
+func (v NullableK3Segment) MarshalJSON() ([]byte, error) {
+	return json.Marshal(v.value)
+}
+
+func (v *NullableK3Segment) UnmarshalJSON(src []byte) error {
+	v.isSet = true
+	return json.Unmarshal(src, &v.value)
+}
+
+
