@@ -1,18 +1,209 @@
 /*
- * METRO2 API
- *
- * Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
+METRO2 API
+
+Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
+
+API version: v1
+*/
+
+// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
 
 package client
 
+import (
+	"encoding/json"
+)
+
 // L1Segment struct for L1Segment
 type L1Segment struct {
-	SegmentIdentifier        string `json:"segmentIdentifier"`
-	ChangeIndicator          int32  `json:"changeIndicator"`
-	NewConsumerAccountNumber string `json:"newConsumerAccountNumber,omitempty"`
-	BalloonPaymentDueDate    string `json:"balloonPaymentDueDate,omitempty"`
+	SegmentIdentifier string `json:"segmentIdentifier"`
+	ChangeIndicator int32 `json:"changeIndicator"`
+	NewConsumerAccountNumber *string `json:"newConsumerAccountNumber,omitempty"`
+	BalloonPaymentDueDate *string `json:"balloonPaymentDueDate,omitempty"`
 }
+
+// NewL1Segment instantiates a new L1Segment object
+// This constructor will assign default values to properties that have it defined,
+// and makes sure properties required by API are set, but the set of arguments
+// will change when the set of required properties is changed
+func NewL1Segment(segmentIdentifier string, changeIndicator int32) *L1Segment {
+	this := L1Segment{}
+	this.SegmentIdentifier = segmentIdentifier
+	this.ChangeIndicator = changeIndicator
+	return &this
+}
+
+// NewL1SegmentWithDefaults instantiates a new L1Segment object
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewL1SegmentWithDefaults() *L1Segment {
+	this := L1Segment{}
+	return &this
+}
+
+// GetSegmentIdentifier returns the SegmentIdentifier field value
+func (o *L1Segment) GetSegmentIdentifier() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.SegmentIdentifier
+}
+
+// GetSegmentIdentifierOk returns a tuple with the SegmentIdentifier field value
+// and a boolean to check if the value has been set.
+func (o *L1Segment) GetSegmentIdentifierOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.SegmentIdentifier, true
+}
+
+// SetSegmentIdentifier sets field value
+func (o *L1Segment) SetSegmentIdentifier(v string) {
+	o.SegmentIdentifier = v
+}
+
+// GetChangeIndicator returns the ChangeIndicator field value
+func (o *L1Segment) GetChangeIndicator() int32 {
+	if o == nil {
+		var ret int32
+		return ret
+	}
+
+	return o.ChangeIndicator
+}
+
+// GetChangeIndicatorOk returns a tuple with the ChangeIndicator field value
+// and a boolean to check if the value has been set.
+func (o *L1Segment) GetChangeIndicatorOk() (*int32, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.ChangeIndicator, true
+}
+
+// SetChangeIndicator sets field value
+func (o *L1Segment) SetChangeIndicator(v int32) {
+	o.ChangeIndicator = v
+}
+
+// GetNewConsumerAccountNumber returns the NewConsumerAccountNumber field value if set, zero value otherwise.
+func (o *L1Segment) GetNewConsumerAccountNumber() string {
+	if o == nil || o.NewConsumerAccountNumber == nil {
+		var ret string
+		return ret
+	}
+	return *o.NewConsumerAccountNumber
+}
+
+// GetNewConsumerAccountNumberOk returns a tuple with the NewConsumerAccountNumber field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *L1Segment) GetNewConsumerAccountNumberOk() (*string, bool) {
+	if o == nil || o.NewConsumerAccountNumber == nil {
+		return nil, false
+	}
+	return o.NewConsumerAccountNumber, true
+}
+
+// HasNewConsumerAccountNumber returns a boolean if a field has been set.
+func (o *L1Segment) HasNewConsumerAccountNumber() bool {
+	if o != nil && o.NewConsumerAccountNumber != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetNewConsumerAccountNumber gets a reference to the given string and assigns it to the NewConsumerAccountNumber field.
+func (o *L1Segment) SetNewConsumerAccountNumber(v string) {
+	o.NewConsumerAccountNumber = &v
+}
+
+// GetBalloonPaymentDueDate returns the BalloonPaymentDueDate field value if set, zero value otherwise.
+func (o *L1Segment) GetBalloonPaymentDueDate() string {
+	if o == nil || o.BalloonPaymentDueDate == nil {
+		var ret string
+		return ret
+	}
+	return *o.BalloonPaymentDueDate
+}
+
+// GetBalloonPaymentDueDateOk returns a tuple with the BalloonPaymentDueDate field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *L1Segment) GetBalloonPaymentDueDateOk() (*string, bool) {
+	if o == nil || o.BalloonPaymentDueDate == nil {
+		return nil, false
+	}
+	return o.BalloonPaymentDueDate, true
+}
+
+// HasBalloonPaymentDueDate returns a boolean if a field has been set.
+func (o *L1Segment) HasBalloonPaymentDueDate() bool {
+	if o != nil && o.BalloonPaymentDueDate != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetBalloonPaymentDueDate gets a reference to the given string and assigns it to the BalloonPaymentDueDate field.
+func (o *L1Segment) SetBalloonPaymentDueDate(v string) {
+	o.BalloonPaymentDueDate = &v
+}
+
+func (o L1Segment) MarshalJSON() ([]byte, error) {
+	toSerialize := map[string]interface{}{}
+	if true {
+		toSerialize["segmentIdentifier"] = o.SegmentIdentifier
+	}
+	if true {
+		toSerialize["changeIndicator"] = o.ChangeIndicator
+	}
+	if o.NewConsumerAccountNumber != nil {
+		toSerialize["newConsumerAccountNumber"] = o.NewConsumerAccountNumber
+	}
+	if o.BalloonPaymentDueDate != nil {
+		toSerialize["balloonPaymentDueDate"] = o.BalloonPaymentDueDate
+	}
+	return json.Marshal(toSerialize)
+}
+
+type NullableL1Segment struct {
+	value *L1Segment
+	isSet bool
+}
+
+func (v NullableL1Segment) Get() *L1Segment {
+	return v.value
+}
+
+func (v *NullableL1Segment) Set(val *L1Segment) {
+	v.value = val
+	v.isSet = true
+}
+
+func (v NullableL1Segment) IsSet() bool {
+	return v.isSet
+}
+
+func (v *NullableL1Segment) Unset() {
+	v.value = nil
+	v.isSet = false
+}
+
+func NewNullableL1Segment(val *L1Segment) *NullableL1Segment {
+	return &NullableL1Segment{value: val, isSet: true}
+}
+
+func (v NullableL1Segment) MarshalJSON() ([]byte, error) {
+	return json.Marshal(v.value)
+}
+
+func (v *NullableL1Segment) UnmarshalJSON(src []byte) error {
+	v.isSet = true
+	return json.Unmarshal(src, &v.value)
+}
+
+
