@@ -1,60 +1,1721 @@
 /*
- * METRO2 API
- *
- * Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
- *
- * API version: v1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
+METRO2 API
+
+Moov Metro2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Metro2 files. Metro2 is an open-source consumer credit history report for credit report file creation and validation.
+
+API version: v1
+*/
+
+// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
 
 package client
 
+import (
+	"encoding/json"
+)
+
 // TrailerRecord struct for TrailerRecord
 type TrailerRecord struct {
-	BlockDescriptorWord              int32  `json:"blockDescriptorWord,omitempty"`
-	RecordDescriptorWord             int32  `json:"recordDescriptorWord"`
-	RecordIdentifier                 string `json:"recordIdentifier"`
-	TotalBaseRecords                 int32  `json:"totalBaseRecords,omitempty"`
-	TotalStatusCodeDF                int32  `json:"totalStatusCodeDF,omitempty"`
-	TotalConsumerSegmentsJ1          int32  `json:"totalConsumerSegmentsJ1,omitempty"`
-	TotalConsumerSegmentsJ2          int32  `json:"totalConsumerSegmentsJ2,omitempty"`
-	BlockCount                       int32  `json:"blockCount,omitempty"`
-	TotalStatusCodeDA                int32  `json:"totalStatusCodeDA,omitempty"`
-	TotalStatusCode05                int32  `json:"totalStatusCode05,omitempty"`
-	TotalStatusCode11                int32  `json:"totalStatusCode11,omitempty"`
-	TotalStatusCode13                int32  `json:"totalStatusCode13,omitempty"`
-	TotalStatusCode61                int32  `json:"totalStatusCode61,omitempty"`
-	TotalStatusCode62                int32  `json:"totalStatusCode62,omitempty"`
-	TotalStatusCode63                int32  `json:"totalStatusCode63,omitempty"`
-	TotalStatusCode64                int32  `json:"totalStatusCode64,omitempty"`
-	TotalStatusCode65                int32  `json:"totalStatusCode65,omitempty"`
-	TotalStatusCode71                int32  `json:"totalStatusCode71,omitempty"`
-	TotalStatusCode78                int32  `json:"totalStatusCode78,omitempty"`
-	TotalStatusCode80                int32  `json:"totalStatusCode80,omitempty"`
-	TotalStatusCode82                int32  `json:"totalStatusCode82,omitempty"`
-	TotalStatusCode83                int32  `json:"totalStatusCode83,omitempty"`
-	TotalStatusCode84                int32  `json:"totalStatusCode84,omitempty"`
-	TotalStatusCode88                int32  `json:"totalStatusCode88,omitempty"`
-	TotalStatusCode89                int32  `json:"totalStatusCode89,omitempty"`
-	TotalStatusCode93                int32  `json:"totalStatusCode93,omitempty"`
-	TotalStatusCode94                int32  `json:"totalStatusCode94,omitempty"`
-	TotalStatusCode95                int32  `json:"totalStatusCode95,omitempty"`
-	TotalStatusCode96                int32  `json:"totalStatusCode96,omitempty"`
-	TotalStatusCode97                int32  `json:"totalStatusCode97,omitempty"`
-	TotalECOACodeZ                   int32  `json:"totalECOACodeZ,omitempty"`
-	TotalEmploymentSegments          int32  `json:"totalEmploymentSegments,omitempty"`
-	TotalOriginalCreditorSegments    int32  `json:"totalOriginalCreditorSegments,omitempty"`
-	TotalPurchasedToSegments         int32  `json:"totalPurchasedToSegments,omitempty"`
-	TotalMortgageInformationSegments int32  `json:"totalMortgageInformationSegments,omitempty"`
-	TotalPaymentInformationSegments  int32  `json:"totalPaymentInformationSegments,omitempty"`
-	TotalChangeSegments              int32  `json:"totalChangeSegments,omitempty"`
-	TotalSocialNumbersAllSegments    int32  `json:"totalSocialNumbersAllSegments,omitempty"`
-	TotalSocialNumbersBaseSegments   int32  `json:"totalSocialNumbersBaseSegments,omitempty"`
-	TotalSocialNumbersJ1Segments     int32  `json:"totalSocialNumbersJ1Segments,omitempty"`
-	TotalSocialNumbersJ2Segments     int32  `json:"totalSocialNumbersJ2Segments,omitempty"`
-	TotalDatesBirthAllSegments       int32  `json:"totalDatesBirthAllSegments,omitempty"`
-	TotalDatesBirthBaseSegments      int32  `json:"totalDatesBirthBaseSegments,omitempty"`
-	TotalDatesBirthJ1Segments        int32  `json:"totalDatesBirthJ1Segments,omitempty"`
-	TotalDatesBirthJ2Segments        int32  `json:"totalDatesBirthJ2Segments,omitempty"`
-	TotalTelephoneNumbersAllSegments int32  `json:"totalTelephoneNumbersAllSegments,omitempty"`
+	BlockDescriptorWord *int32 `json:"blockDescriptorWord,omitempty"`
+	RecordDescriptorWord int32 `json:"recordDescriptorWord"`
+	RecordIdentifier string `json:"recordIdentifier"`
+	TotalBaseRecords *int32 `json:"totalBaseRecords,omitempty"`
+	TotalStatusCodeDF *int32 `json:"totalStatusCodeDF,omitempty"`
+	TotalConsumerSegmentsJ1 *int32 `json:"totalConsumerSegmentsJ1,omitempty"`
+	TotalConsumerSegmentsJ2 *int32 `json:"totalConsumerSegmentsJ2,omitempty"`
+	BlockCount *int32 `json:"blockCount,omitempty"`
+	TotalStatusCodeDA *int32 `json:"totalStatusCodeDA,omitempty"`
+	TotalStatusCode05 *int32 `json:"totalStatusCode05,omitempty"`
+	TotalStatusCode11 *int32 `json:"totalStatusCode11,omitempty"`
+	TotalStatusCode13 *int32 `json:"totalStatusCode13,omitempty"`
+	TotalStatusCode61 *int32 `json:"totalStatusCode61,omitempty"`
+	TotalStatusCode62 *int32 `json:"totalStatusCode62,omitempty"`
+	TotalStatusCode63 *int32 `json:"totalStatusCode63,omitempty"`
+	TotalStatusCode64 *int32 `json:"totalStatusCode64,omitempty"`
+	TotalStatusCode65 *int32 `json:"totalStatusCode65,omitempty"`
+	TotalStatusCode71 *int32 `json:"totalStatusCode71,omitempty"`
+	TotalStatusCode78 *int32 `json:"totalStatusCode78,omitempty"`
+	TotalStatusCode80 *int32 `json:"totalStatusCode80,omitempty"`
+	TotalStatusCode82 *int32 `json:"totalStatusCode82,omitempty"`
+	TotalStatusCode83 *int32 `json:"totalStatusCode83,omitempty"`
+	TotalStatusCode84 *int32 `json:"totalStatusCode84,omitempty"`
+	TotalStatusCode88 *int32 `json:"totalStatusCode88,omitempty"`
+	TotalStatusCode89 *int32 `json:"totalStatusCode89,omitempty"`
+	TotalStatusCode93 *int32 `json:"totalStatusCode93,omitempty"`
+	TotalStatusCode94 *int32 `json:"totalStatusCode94,omitempty"`
+	TotalStatusCode95 *int32 `json:"totalStatusCode95,omitempty"`
+	TotalStatusCode96 *int32 `json:"totalStatusCode96,omitempty"`
+	TotalStatusCode97 *int32 `json:"totalStatusCode97,omitempty"`
+	TotalECOACodeZ *int32 `json:"totalECOACodeZ,omitempty"`
+	TotalEmploymentSegments *int32 `json:"totalEmploymentSegments,omitempty"`
+	TotalOriginalCreditorSegments *int32 `json:"totalOriginalCreditorSegments,omitempty"`
+	TotalPurchasedToSegments *int32 `json:"totalPurchasedToSegments,omitempty"`
+	TotalMortgageInformationSegments *int32 `json:"totalMortgageInformationSegments,omitempty"`
+	TotalPaymentInformationSegments *int32 `json:"totalPaymentInformationSegments,omitempty"`
+	TotalChangeSegments *int32 `json:"totalChangeSegments,omitempty"`
+	TotalSocialNumbersAllSegments *int32 `json:"totalSocialNumbersAllSegments,omitempty"`
+	TotalSocialNumbersBaseSegments *int32 `json:"totalSocialNumbersBaseSegments,omitempty"`
+	TotalSocialNumbersJ1Segments *int32 `json:"totalSocialNumbersJ1Segments,omitempty"`
+	TotalSocialNumbersJ2Segments *int32 `json:"totalSocialNumbersJ2Segments,omitempty"`
+	TotalDatesBirthAllSegments *int32 `json:"totalDatesBirthAllSegments,omitempty"`
+	TotalDatesBirthBaseSegments *int32 `json:"totalDatesBirthBaseSegments,omitempty"`
+	TotalDatesBirthJ1Segments *int32 `json:"totalDatesBirthJ1Segments,omitempty"`
+	TotalDatesBirthJ2Segments *int32 `json:"totalDatesBirthJ2Segments,omitempty"`
+	TotalTelephoneNumbersAllSegments *int32 `json:"totalTelephoneNumbersAllSegments,omitempty"`
 }
+
+// NewTrailerRecord instantiates a new TrailerRecord object
+// This constructor will assign default values to properties that have it defined,
+// and makes sure properties required by API are set, but the set of arguments
+// will change when the set of required properties is changed
+func NewTrailerRecord(recordDescriptorWord int32, recordIdentifier string) *TrailerRecord {
+	this := TrailerRecord{}
+	this.RecordDescriptorWord = recordDescriptorWord
+	this.RecordIdentifier = recordIdentifier
+	return &this
+}
+
+// NewTrailerRecordWithDefaults instantiates a new TrailerRecord object
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewTrailerRecordWithDefaults() *TrailerRecord {
+	this := TrailerRecord{}
+	return &this
+}
+
+// GetBlockDescriptorWord returns the BlockDescriptorWord field value if set, zero value otherwise.
+func (o *TrailerRecord) GetBlockDescriptorWord() int32 {
+	if o == nil || o.BlockDescriptorWord == nil {
+		var ret int32
+		return ret
+	}
+	return *o.BlockDescriptorWord
+}
+
+// GetBlockDescriptorWordOk returns a tuple with the BlockDescriptorWord field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetBlockDescriptorWordOk() (*int32, bool) {
+	if o == nil || o.BlockDescriptorWord == nil {
+		return nil, false
+	}
+	return o.BlockDescriptorWord, true
+}
+
+// HasBlockDescriptorWord returns a boolean if a field has been set.
+func (o *TrailerRecord) HasBlockDescriptorWord() bool {
+	if o != nil && o.BlockDescriptorWord != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetBlockDescriptorWord gets a reference to the given int32 and assigns it to the BlockDescriptorWord field.
+func (o *TrailerRecord) SetBlockDescriptorWord(v int32) {
+	o.BlockDescriptorWord = &v
+}
+
+// GetRecordDescriptorWord returns the RecordDescriptorWord field value
+func (o *TrailerRecord) GetRecordDescriptorWord() int32 {
+	if o == nil {
+		var ret int32
+		return ret
+	}
+
+	return o.RecordDescriptorWord
+}
+
+// GetRecordDescriptorWordOk returns a tuple with the RecordDescriptorWord field value
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetRecordDescriptorWordOk() (*int32, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.RecordDescriptorWord, true
+}
+
+// SetRecordDescriptorWord sets field value
+func (o *TrailerRecord) SetRecordDescriptorWord(v int32) {
+	o.RecordDescriptorWord = v
+}
+
+// GetRecordIdentifier returns the RecordIdentifier field value
+func (o *TrailerRecord) GetRecordIdentifier() string {
+	if o == nil {
+		var ret string
+		return ret
+	}
+
+	return o.RecordIdentifier
+}
+
+// GetRecordIdentifierOk returns a tuple with the RecordIdentifier field value
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetRecordIdentifierOk() (*string, bool) {
+	if o == nil  {
+		return nil, false
+	}
+	return &o.RecordIdentifier, true
+}
+
+// SetRecordIdentifier sets field value
+func (o *TrailerRecord) SetRecordIdentifier(v string) {
+	o.RecordIdentifier = v
+}
+
+// GetTotalBaseRecords returns the TotalBaseRecords field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalBaseRecords() int32 {
+	if o == nil || o.TotalBaseRecords == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalBaseRecords
+}
+
+// GetTotalBaseRecordsOk returns a tuple with the TotalBaseRecords field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalBaseRecordsOk() (*int32, bool) {
+	if o == nil || o.TotalBaseRecords == nil {
+		return nil, false
+	}
+	return o.TotalBaseRecords, true
+}
+
+// HasTotalBaseRecords returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalBaseRecords() bool {
+	if o != nil && o.TotalBaseRecords != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalBaseRecords gets a reference to the given int32 and assigns it to the TotalBaseRecords field.
+func (o *TrailerRecord) SetTotalBaseRecords(v int32) {
+	o.TotalBaseRecords = &v
+}
+
+// GetTotalStatusCodeDF returns the TotalStatusCodeDF field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCodeDF() int32 {
+	if o == nil || o.TotalStatusCodeDF == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCodeDF
+}
+
+// GetTotalStatusCodeDFOk returns a tuple with the TotalStatusCodeDF field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCodeDFOk() (*int32, bool) {
+	if o == nil || o.TotalStatusCodeDF == nil {
+		return nil, false
+	}
+	return o.TotalStatusCodeDF, true
+}
+
+// HasTotalStatusCodeDF returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCodeDF() bool {
+	if o != nil && o.TotalStatusCodeDF != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCodeDF gets a reference to the given int32 and assigns it to the TotalStatusCodeDF field.
+func (o *TrailerRecord) SetTotalStatusCodeDF(v int32) {
+	o.TotalStatusCodeDF = &v
+}
+
+// GetTotalConsumerSegmentsJ1 returns the TotalConsumerSegmentsJ1 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalConsumerSegmentsJ1() int32 {
+	if o == nil || o.TotalConsumerSegmentsJ1 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalConsumerSegmentsJ1
+}
+
+// GetTotalConsumerSegmentsJ1Ok returns a tuple with the TotalConsumerSegmentsJ1 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalConsumerSegmentsJ1Ok() (*int32, bool) {
+	if o == nil || o.TotalConsumerSegmentsJ1 == nil {
+		return nil, false
+	}
+	return o.TotalConsumerSegmentsJ1, true
+}
+
+// HasTotalConsumerSegmentsJ1 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalConsumerSegmentsJ1() bool {
+	if o != nil && o.TotalConsumerSegmentsJ1 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalConsumerSegmentsJ1 gets a reference to the given int32 and assigns it to the TotalConsumerSegmentsJ1 field.
+func (o *TrailerRecord) SetTotalConsumerSegmentsJ1(v int32) {
+	o.TotalConsumerSegmentsJ1 = &v
+}
+
+// GetTotalConsumerSegmentsJ2 returns the TotalConsumerSegmentsJ2 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalConsumerSegmentsJ2() int32 {
+	if o == nil || o.TotalConsumerSegmentsJ2 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalConsumerSegmentsJ2
+}
+
+// GetTotalConsumerSegmentsJ2Ok returns a tuple with the TotalConsumerSegmentsJ2 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalConsumerSegmentsJ2Ok() (*int32, bool) {
+	if o == nil || o.TotalConsumerSegmentsJ2 == nil {
+		return nil, false
+	}
+	return o.TotalConsumerSegmentsJ2, true
+}
+
+// HasTotalConsumerSegmentsJ2 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalConsumerSegmentsJ2() bool {
+	if o != nil && o.TotalConsumerSegmentsJ2 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalConsumerSegmentsJ2 gets a reference to the given int32 and assigns it to the TotalConsumerSegmentsJ2 field.
+func (o *TrailerRecord) SetTotalConsumerSegmentsJ2(v int32) {
+	o.TotalConsumerSegmentsJ2 = &v
+}
+
+// GetBlockCount returns the BlockCount field value if set, zero value otherwise.
+func (o *TrailerRecord) GetBlockCount() int32 {
+	if o == nil || o.BlockCount == nil {
+		var ret int32
+		return ret
+	}
+	return *o.BlockCount
+}
+
+// GetBlockCountOk returns a tuple with the BlockCount field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetBlockCountOk() (*int32, bool) {
+	if o == nil || o.BlockCount == nil {
+		return nil, false
+	}
+	return o.BlockCount, true
+}
+
+// HasBlockCount returns a boolean if a field has been set.
+func (o *TrailerRecord) HasBlockCount() bool {
+	if o != nil && o.BlockCount != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetBlockCount gets a reference to the given int32 and assigns it to the BlockCount field.
+func (o *TrailerRecord) SetBlockCount(v int32) {
+	o.BlockCount = &v
+}
+
+// GetTotalStatusCodeDA returns the TotalStatusCodeDA field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCodeDA() int32 {
+	if o == nil || o.TotalStatusCodeDA == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCodeDA
+}
+
+// GetTotalStatusCodeDAOk returns a tuple with the TotalStatusCodeDA field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCodeDAOk() (*int32, bool) {
+	if o == nil || o.TotalStatusCodeDA == nil {
+		return nil, false
+	}
+	return o.TotalStatusCodeDA, true
+}
+
+// HasTotalStatusCodeDA returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCodeDA() bool {
+	if o != nil && o.TotalStatusCodeDA != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCodeDA gets a reference to the given int32 and assigns it to the TotalStatusCodeDA field.
+func (o *TrailerRecord) SetTotalStatusCodeDA(v int32) {
+	o.TotalStatusCodeDA = &v
+}
+
+// GetTotalStatusCode05 returns the TotalStatusCode05 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode05() int32 {
+	if o == nil || o.TotalStatusCode05 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode05
+}
+
+// GetTotalStatusCode05Ok returns a tuple with the TotalStatusCode05 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode05Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode05 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode05, true
+}
+
+// HasTotalStatusCode05 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode05() bool {
+	if o != nil && o.TotalStatusCode05 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode05 gets a reference to the given int32 and assigns it to the TotalStatusCode05 field.
+func (o *TrailerRecord) SetTotalStatusCode05(v int32) {
+	o.TotalStatusCode05 = &v
+}
+
+// GetTotalStatusCode11 returns the TotalStatusCode11 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode11() int32 {
+	if o == nil || o.TotalStatusCode11 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode11
+}
+
+// GetTotalStatusCode11Ok returns a tuple with the TotalStatusCode11 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode11Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode11 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode11, true
+}
+
+// HasTotalStatusCode11 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode11() bool {
+	if o != nil && o.TotalStatusCode11 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode11 gets a reference to the given int32 and assigns it to the TotalStatusCode11 field.
+func (o *TrailerRecord) SetTotalStatusCode11(v int32) {
+	o.TotalStatusCode11 = &v
+}
+
+// GetTotalStatusCode13 returns the TotalStatusCode13 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode13() int32 {
+	if o == nil || o.TotalStatusCode13 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode13
+}
+
+// GetTotalStatusCode13Ok returns a tuple with the TotalStatusCode13 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode13Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode13 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode13, true
+}
+
+// HasTotalStatusCode13 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode13() bool {
+	if o != nil && o.TotalStatusCode13 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode13 gets a reference to the given int32 and assigns it to the TotalStatusCode13 field.
+func (o *TrailerRecord) SetTotalStatusCode13(v int32) {
+	o.TotalStatusCode13 = &v
+}
+
+// GetTotalStatusCode61 returns the TotalStatusCode61 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode61() int32 {
+	if o == nil || o.TotalStatusCode61 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode61
+}
+
+// GetTotalStatusCode61Ok returns a tuple with the TotalStatusCode61 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode61Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode61 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode61, true
+}
+
+// HasTotalStatusCode61 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode61() bool {
+	if o != nil && o.TotalStatusCode61 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode61 gets a reference to the given int32 and assigns it to the TotalStatusCode61 field.
+func (o *TrailerRecord) SetTotalStatusCode61(v int32) {
+	o.TotalStatusCode61 = &v
+}
+
+// GetTotalStatusCode62 returns the TotalStatusCode62 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode62() int32 {
+	if o == nil || o.TotalStatusCode62 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode62
+}
+
+// GetTotalStatusCode62Ok returns a tuple with the TotalStatusCode62 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode62Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode62 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode62, true
+}
+
+// HasTotalStatusCode62 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode62() bool {
+	if o != nil && o.TotalStatusCode62 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode62 gets a reference to the given int32 and assigns it to the TotalStatusCode62 field.
+func (o *TrailerRecord) SetTotalStatusCode62(v int32) {
+	o.TotalStatusCode62 = &v
+}
+
+// GetTotalStatusCode63 returns the TotalStatusCode63 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode63() int32 {
+	if o == nil || o.TotalStatusCode63 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode63
+}
+
+// GetTotalStatusCode63Ok returns a tuple with the TotalStatusCode63 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode63Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode63 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode63, true
+}
+
+// HasTotalStatusCode63 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode63() bool {
+	if o != nil && o.TotalStatusCode63 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode63 gets a reference to the given int32 and assigns it to the TotalStatusCode63 field.
+func (o *TrailerRecord) SetTotalStatusCode63(v int32) {
+	o.TotalStatusCode63 = &v
+}
+
+// GetTotalStatusCode64 returns the TotalStatusCode64 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode64() int32 {
+	if o == nil || o.TotalStatusCode64 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode64
+}
+
+// GetTotalStatusCode64Ok returns a tuple with the TotalStatusCode64 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode64Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode64 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode64, true
+}
+
+// HasTotalStatusCode64 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode64() bool {
+	if o != nil && o.TotalStatusCode64 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode64 gets a reference to the given int32 and assigns it to the TotalStatusCode64 field.
+func (o *TrailerRecord) SetTotalStatusCode64(v int32) {
+	o.TotalStatusCode64 = &v
+}
+
+// GetTotalStatusCode65 returns the TotalStatusCode65 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode65() int32 {
+	if o == nil || o.TotalStatusCode65 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode65
+}
+
+// GetTotalStatusCode65Ok returns a tuple with the TotalStatusCode65 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode65Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode65 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode65, true
+}
+
+// HasTotalStatusCode65 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode65() bool {
+	if o != nil && o.TotalStatusCode65 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode65 gets a reference to the given int32 and assigns it to the TotalStatusCode65 field.
+func (o *TrailerRecord) SetTotalStatusCode65(v int32) {
+	o.TotalStatusCode65 = &v
+}
+
+// GetTotalStatusCode71 returns the TotalStatusCode71 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode71() int32 {
+	if o == nil || o.TotalStatusCode71 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode71
+}
+
+// GetTotalStatusCode71Ok returns a tuple with the TotalStatusCode71 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode71Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode71 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode71, true
+}
+
+// HasTotalStatusCode71 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode71() bool {
+	if o != nil && o.TotalStatusCode71 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode71 gets a reference to the given int32 and assigns it to the TotalStatusCode71 field.
+func (o *TrailerRecord) SetTotalStatusCode71(v int32) {
+	o.TotalStatusCode71 = &v
+}
+
+// GetTotalStatusCode78 returns the TotalStatusCode78 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode78() int32 {
+	if o == nil || o.TotalStatusCode78 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode78
+}
+
+// GetTotalStatusCode78Ok returns a tuple with the TotalStatusCode78 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode78Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode78 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode78, true
+}
+
+// HasTotalStatusCode78 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode78() bool {
+	if o != nil && o.TotalStatusCode78 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode78 gets a reference to the given int32 and assigns it to the TotalStatusCode78 field.
+func (o *TrailerRecord) SetTotalStatusCode78(v int32) {
+	o.TotalStatusCode78 = &v
+}
+
+// GetTotalStatusCode80 returns the TotalStatusCode80 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode80() int32 {
+	if o == nil || o.TotalStatusCode80 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode80
+}
+
+// GetTotalStatusCode80Ok returns a tuple with the TotalStatusCode80 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode80Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode80 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode80, true
+}
+
+// HasTotalStatusCode80 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode80() bool {
+	if o != nil && o.TotalStatusCode80 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode80 gets a reference to the given int32 and assigns it to the TotalStatusCode80 field.
+func (o *TrailerRecord) SetTotalStatusCode80(v int32) {
+	o.TotalStatusCode80 = &v
+}
+
+// GetTotalStatusCode82 returns the TotalStatusCode82 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode82() int32 {
+	if o == nil || o.TotalStatusCode82 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode82
+}
+
+// GetTotalStatusCode82Ok returns a tuple with the TotalStatusCode82 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode82Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode82 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode82, true
+}
+
+// HasTotalStatusCode82 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode82() bool {
+	if o != nil && o.TotalStatusCode82 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode82 gets a reference to the given int32 and assigns it to the TotalStatusCode82 field.
+func (o *TrailerRecord) SetTotalStatusCode82(v int32) {
+	o.TotalStatusCode82 = &v
+}
+
+// GetTotalStatusCode83 returns the TotalStatusCode83 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode83() int32 {
+	if o == nil || o.TotalStatusCode83 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode83
+}
+
+// GetTotalStatusCode83Ok returns a tuple with the TotalStatusCode83 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode83Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode83 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode83, true
+}
+
+// HasTotalStatusCode83 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode83() bool {
+	if o != nil && o.TotalStatusCode83 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode83 gets a reference to the given int32 and assigns it to the TotalStatusCode83 field.
+func (o *TrailerRecord) SetTotalStatusCode83(v int32) {
+	o.TotalStatusCode83 = &v
+}
+
+// GetTotalStatusCode84 returns the TotalStatusCode84 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode84() int32 {
+	if o == nil || o.TotalStatusCode84 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode84
+}
+
+// GetTotalStatusCode84Ok returns a tuple with the TotalStatusCode84 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode84Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode84 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode84, true
+}
+
+// HasTotalStatusCode84 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode84() bool {
+	if o != nil && o.TotalStatusCode84 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode84 gets a reference to the given int32 and assigns it to the TotalStatusCode84 field.
+func (o *TrailerRecord) SetTotalStatusCode84(v int32) {
+	o.TotalStatusCode84 = &v
+}
+
+// GetTotalStatusCode88 returns the TotalStatusCode88 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode88() int32 {
+	if o == nil || o.TotalStatusCode88 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode88
+}
+
+// GetTotalStatusCode88Ok returns a tuple with the TotalStatusCode88 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode88Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode88 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode88, true
+}
+
+// HasTotalStatusCode88 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode88() bool {
+	if o != nil && o.TotalStatusCode88 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode88 gets a reference to the given int32 and assigns it to the TotalStatusCode88 field.
+func (o *TrailerRecord) SetTotalStatusCode88(v int32) {
+	o.TotalStatusCode88 = &v
+}
+
+// GetTotalStatusCode89 returns the TotalStatusCode89 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode89() int32 {
+	if o == nil || o.TotalStatusCode89 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode89
+}
+
+// GetTotalStatusCode89Ok returns a tuple with the TotalStatusCode89 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode89Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode89 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode89, true
+}
+
+// HasTotalStatusCode89 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode89() bool {
+	if o != nil && o.TotalStatusCode89 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode89 gets a reference to the given int32 and assigns it to the TotalStatusCode89 field.
+func (o *TrailerRecord) SetTotalStatusCode89(v int32) {
+	o.TotalStatusCode89 = &v
+}
+
+// GetTotalStatusCode93 returns the TotalStatusCode93 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode93() int32 {
+	if o == nil || o.TotalStatusCode93 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode93
+}
+
+// GetTotalStatusCode93Ok returns a tuple with the TotalStatusCode93 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode93Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode93 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode93, true
+}
+
+// HasTotalStatusCode93 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode93() bool {
+	if o != nil && o.TotalStatusCode93 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode93 gets a reference to the given int32 and assigns it to the TotalStatusCode93 field.
+func (o *TrailerRecord) SetTotalStatusCode93(v int32) {
+	o.TotalStatusCode93 = &v
+}
+
+// GetTotalStatusCode94 returns the TotalStatusCode94 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode94() int32 {
+	if o == nil || o.TotalStatusCode94 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode94
+}
+
+// GetTotalStatusCode94Ok returns a tuple with the TotalStatusCode94 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode94Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode94 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode94, true
+}
+
+// HasTotalStatusCode94 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode94() bool {
+	if o != nil && o.TotalStatusCode94 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode94 gets a reference to the given int32 and assigns it to the TotalStatusCode94 field.
+func (o *TrailerRecord) SetTotalStatusCode94(v int32) {
+	o.TotalStatusCode94 = &v
+}
+
+// GetTotalStatusCode95 returns the TotalStatusCode95 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode95() int32 {
+	if o == nil || o.TotalStatusCode95 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode95
+}
+
+// GetTotalStatusCode95Ok returns a tuple with the TotalStatusCode95 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode95Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode95 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode95, true
+}
+
+// HasTotalStatusCode95 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode95() bool {
+	if o != nil && o.TotalStatusCode95 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode95 gets a reference to the given int32 and assigns it to the TotalStatusCode95 field.
+func (o *TrailerRecord) SetTotalStatusCode95(v int32) {
+	o.TotalStatusCode95 = &v
+}
+
+// GetTotalStatusCode96 returns the TotalStatusCode96 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode96() int32 {
+	if o == nil || o.TotalStatusCode96 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode96
+}
+
+// GetTotalStatusCode96Ok returns a tuple with the TotalStatusCode96 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode96Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode96 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode96, true
+}
+
+// HasTotalStatusCode96 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode96() bool {
+	if o != nil && o.TotalStatusCode96 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode96 gets a reference to the given int32 and assigns it to the TotalStatusCode96 field.
+func (o *TrailerRecord) SetTotalStatusCode96(v int32) {
+	o.TotalStatusCode96 = &v
+}
+
+// GetTotalStatusCode97 returns the TotalStatusCode97 field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalStatusCode97() int32 {
+	if o == nil || o.TotalStatusCode97 == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalStatusCode97
+}
+
+// GetTotalStatusCode97Ok returns a tuple with the TotalStatusCode97 field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalStatusCode97Ok() (*int32, bool) {
+	if o == nil || o.TotalStatusCode97 == nil {
+		return nil, false
+	}
+	return o.TotalStatusCode97, true
+}
+
+// HasTotalStatusCode97 returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalStatusCode97() bool {
+	if o != nil && o.TotalStatusCode97 != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalStatusCode97 gets a reference to the given int32 and assigns it to the TotalStatusCode97 field.
+func (o *TrailerRecord) SetTotalStatusCode97(v int32) {
+	o.TotalStatusCode97 = &v
+}
+
+// GetTotalECOACodeZ returns the TotalECOACodeZ field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalECOACodeZ() int32 {
+	if o == nil || o.TotalECOACodeZ == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalECOACodeZ
+}
+
+// GetTotalECOACodeZOk returns a tuple with the TotalECOACodeZ field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalECOACodeZOk() (*int32, bool) {
+	if o == nil || o.TotalECOACodeZ == nil {
+		return nil, false
+	}
+	return o.TotalECOACodeZ, true
+}
+
+// HasTotalECOACodeZ returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalECOACodeZ() bool {
+	if o != nil && o.TotalECOACodeZ != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalECOACodeZ gets a reference to the given int32 and assigns it to the TotalECOACodeZ field.
+func (o *TrailerRecord) SetTotalECOACodeZ(v int32) {
+	o.TotalECOACodeZ = &v
+}
+
+// GetTotalEmploymentSegments returns the TotalEmploymentSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalEmploymentSegments() int32 {
+	if o == nil || o.TotalEmploymentSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalEmploymentSegments
+}
+
+// GetTotalEmploymentSegmentsOk returns a tuple with the TotalEmploymentSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalEmploymentSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalEmploymentSegments == nil {
+		return nil, false
+	}
+	return o.TotalEmploymentSegments, true
+}
+
+// HasTotalEmploymentSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalEmploymentSegments() bool {
+	if o != nil && o.TotalEmploymentSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalEmploymentSegments gets a reference to the given int32 and assigns it to the TotalEmploymentSegments field.
+func (o *TrailerRecord) SetTotalEmploymentSegments(v int32) {
+	o.TotalEmploymentSegments = &v
+}
+
+// GetTotalOriginalCreditorSegments returns the TotalOriginalCreditorSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalOriginalCreditorSegments() int32 {
+	if o == nil || o.TotalOriginalCreditorSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalOriginalCreditorSegments
+}
+
+// GetTotalOriginalCreditorSegmentsOk returns a tuple with the TotalOriginalCreditorSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalOriginalCreditorSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalOriginalCreditorSegments == nil {
+		return nil, false
+	}
+	return o.TotalOriginalCreditorSegments, true
+}
+
+// HasTotalOriginalCreditorSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalOriginalCreditorSegments() bool {
+	if o != nil && o.TotalOriginalCreditorSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalOriginalCreditorSegments gets a reference to the given int32 and assigns it to the TotalOriginalCreditorSegments field.
+func (o *TrailerRecord) SetTotalOriginalCreditorSegments(v int32) {
+	o.TotalOriginalCreditorSegments = &v
+}
+
+// GetTotalPurchasedToSegments returns the TotalPurchasedToSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalPurchasedToSegments() int32 {
+	if o == nil || o.TotalPurchasedToSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalPurchasedToSegments
+}
+
+// GetTotalPurchasedToSegmentsOk returns a tuple with the TotalPurchasedToSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalPurchasedToSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalPurchasedToSegments == nil {
+		return nil, false
+	}
+	return o.TotalPurchasedToSegments, true
+}
+
+// HasTotalPurchasedToSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalPurchasedToSegments() bool {
+	if o != nil && o.TotalPurchasedToSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalPurchasedToSegments gets a reference to the given int32 and assigns it to the TotalPurchasedToSegments field.
+func (o *TrailerRecord) SetTotalPurchasedToSegments(v int32) {
+	o.TotalPurchasedToSegments = &v
+}
+
+// GetTotalMortgageInformationSegments returns the TotalMortgageInformationSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalMortgageInformationSegments() int32 {
+	if o == nil || o.TotalMortgageInformationSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalMortgageInformationSegments
+}
+
+// GetTotalMortgageInformationSegmentsOk returns a tuple with the TotalMortgageInformationSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalMortgageInformationSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalMortgageInformationSegments == nil {
+		return nil, false
+	}
+	return o.TotalMortgageInformationSegments, true
+}
+
+// HasTotalMortgageInformationSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalMortgageInformationSegments() bool {
+	if o != nil && o.TotalMortgageInformationSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalMortgageInformationSegments gets a reference to the given int32 and assigns it to the TotalMortgageInformationSegments field.
+func (o *TrailerRecord) SetTotalMortgageInformationSegments(v int32) {
+	o.TotalMortgageInformationSegments = &v
+}
+
+// GetTotalPaymentInformationSegments returns the TotalPaymentInformationSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalPaymentInformationSegments() int32 {
+	if o == nil || o.TotalPaymentInformationSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalPaymentInformationSegments
+}
+
+// GetTotalPaymentInformationSegmentsOk returns a tuple with the TotalPaymentInformationSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalPaymentInformationSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalPaymentInformationSegments == nil {
+		return nil, false
+	}
+	return o.TotalPaymentInformationSegments, true
+}
+
+// HasTotalPaymentInformationSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalPaymentInformationSegments() bool {
+	if o != nil && o.TotalPaymentInformationSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalPaymentInformationSegments gets a reference to the given int32 and assigns it to the TotalPaymentInformationSegments field.
+func (o *TrailerRecord) SetTotalPaymentInformationSegments(v int32) {
+	o.TotalPaymentInformationSegments = &v
+}
+
+// GetTotalChangeSegments returns the TotalChangeSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalChangeSegments() int32 {
+	if o == nil || o.TotalChangeSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalChangeSegments
+}
+
+// GetTotalChangeSegmentsOk returns a tuple with the TotalChangeSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalChangeSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalChangeSegments == nil {
+		return nil, false
+	}
+	return o.TotalChangeSegments, true
+}
+
+// HasTotalChangeSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalChangeSegments() bool {
+	if o != nil && o.TotalChangeSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalChangeSegments gets a reference to the given int32 and assigns it to the TotalChangeSegments field.
+func (o *TrailerRecord) SetTotalChangeSegments(v int32) {
+	o.TotalChangeSegments = &v
+}
+
+// GetTotalSocialNumbersAllSegments returns the TotalSocialNumbersAllSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalSocialNumbersAllSegments() int32 {
+	if o == nil || o.TotalSocialNumbersAllSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalSocialNumbersAllSegments
+}
+
+// GetTotalSocialNumbersAllSegmentsOk returns a tuple with the TotalSocialNumbersAllSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalSocialNumbersAllSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalSocialNumbersAllSegments == nil {
+		return nil, false
+	}
+	return o.TotalSocialNumbersAllSegments, true
+}
+
+// HasTotalSocialNumbersAllSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalSocialNumbersAllSegments() bool {
+	if o != nil && o.TotalSocialNumbersAllSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalSocialNumbersAllSegments gets a reference to the given int32 and assigns it to the TotalSocialNumbersAllSegments field.
+func (o *TrailerRecord) SetTotalSocialNumbersAllSegments(v int32) {
+	o.TotalSocialNumbersAllSegments = &v
+}
+
+// GetTotalSocialNumbersBaseSegments returns the TotalSocialNumbersBaseSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalSocialNumbersBaseSegments() int32 {
+	if o == nil || o.TotalSocialNumbersBaseSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalSocialNumbersBaseSegments
+}
+
+// GetTotalSocialNumbersBaseSegmentsOk returns a tuple with the TotalSocialNumbersBaseSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalSocialNumbersBaseSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalSocialNumbersBaseSegments == nil {
+		return nil, false
+	}
+	return o.TotalSocialNumbersBaseSegments, true
+}
+
+// HasTotalSocialNumbersBaseSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalSocialNumbersBaseSegments() bool {
+	if o != nil && o.TotalSocialNumbersBaseSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalSocialNumbersBaseSegments gets a reference to the given int32 and assigns it to the TotalSocialNumbersBaseSegments field.
+func (o *TrailerRecord) SetTotalSocialNumbersBaseSegments(v int32) {
+	o.TotalSocialNumbersBaseSegments = &v
+}
+
+// GetTotalSocialNumbersJ1Segments returns the TotalSocialNumbersJ1Segments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalSocialNumbersJ1Segments() int32 {
+	if o == nil || o.TotalSocialNumbersJ1Segments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalSocialNumbersJ1Segments
+}
+
+// GetTotalSocialNumbersJ1SegmentsOk returns a tuple with the TotalSocialNumbersJ1Segments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalSocialNumbersJ1SegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalSocialNumbersJ1Segments == nil {
+		return nil, false
+	}
+	return o.TotalSocialNumbersJ1Segments, true
+}
+
+// HasTotalSocialNumbersJ1Segments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalSocialNumbersJ1Segments() bool {
+	if o != nil && o.TotalSocialNumbersJ1Segments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalSocialNumbersJ1Segments gets a reference to the given int32 and assigns it to the TotalSocialNumbersJ1Segments field.
+func (o *TrailerRecord) SetTotalSocialNumbersJ1Segments(v int32) {
+	o.TotalSocialNumbersJ1Segments = &v
+}
+
+// GetTotalSocialNumbersJ2Segments returns the TotalSocialNumbersJ2Segments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalSocialNumbersJ2Segments() int32 {
+	if o == nil || o.TotalSocialNumbersJ2Segments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalSocialNumbersJ2Segments
+}
+
+// GetTotalSocialNumbersJ2SegmentsOk returns a tuple with the TotalSocialNumbersJ2Segments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalSocialNumbersJ2SegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalSocialNumbersJ2Segments == nil {
+		return nil, false
+	}
+	return o.TotalSocialNumbersJ2Segments, true
+}
+
+// HasTotalSocialNumbersJ2Segments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalSocialNumbersJ2Segments() bool {
+	if o != nil && o.TotalSocialNumbersJ2Segments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalSocialNumbersJ2Segments gets a reference to the given int32 and assigns it to the TotalSocialNumbersJ2Segments field.
+func (o *TrailerRecord) SetTotalSocialNumbersJ2Segments(v int32) {
+	o.TotalSocialNumbersJ2Segments = &v
+}
+
+// GetTotalDatesBirthAllSegments returns the TotalDatesBirthAllSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalDatesBirthAllSegments() int32 {
+	if o == nil || o.TotalDatesBirthAllSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalDatesBirthAllSegments
+}
+
+// GetTotalDatesBirthAllSegmentsOk returns a tuple with the TotalDatesBirthAllSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalDatesBirthAllSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalDatesBirthAllSegments == nil {
+		return nil, false
+	}
+	return o.TotalDatesBirthAllSegments, true
+}
+
+// HasTotalDatesBirthAllSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalDatesBirthAllSegments() bool {
+	if o != nil && o.TotalDatesBirthAllSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalDatesBirthAllSegments gets a reference to the given int32 and assigns it to the TotalDatesBirthAllSegments field.
+func (o *TrailerRecord) SetTotalDatesBirthAllSegments(v int32) {
+	o.TotalDatesBirthAllSegments = &v
+}
+
+// GetTotalDatesBirthBaseSegments returns the TotalDatesBirthBaseSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalDatesBirthBaseSegments() int32 {
+	if o == nil || o.TotalDatesBirthBaseSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalDatesBirthBaseSegments
+}
+
+// GetTotalDatesBirthBaseSegmentsOk returns a tuple with the TotalDatesBirthBaseSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalDatesBirthBaseSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalDatesBirthBaseSegments == nil {
+		return nil, false
+	}
+	return o.TotalDatesBirthBaseSegments, true
+}
+
+// HasTotalDatesBirthBaseSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalDatesBirthBaseSegments() bool {
+	if o != nil && o.TotalDatesBirthBaseSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalDatesBirthBaseSegments gets a reference to the given int32 and assigns it to the TotalDatesBirthBaseSegments field.
+func (o *TrailerRecord) SetTotalDatesBirthBaseSegments(v int32) {
+	o.TotalDatesBirthBaseSegments = &v
+}
+
+// GetTotalDatesBirthJ1Segments returns the TotalDatesBirthJ1Segments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalDatesBirthJ1Segments() int32 {
+	if o == nil || o.TotalDatesBirthJ1Segments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalDatesBirthJ1Segments
+}
+
+// GetTotalDatesBirthJ1SegmentsOk returns a tuple with the TotalDatesBirthJ1Segments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalDatesBirthJ1SegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalDatesBirthJ1Segments == nil {
+		return nil, false
+	}
+	return o.TotalDatesBirthJ1Segments, true
+}
+
+// HasTotalDatesBirthJ1Segments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalDatesBirthJ1Segments() bool {
+	if o != nil && o.TotalDatesBirthJ1Segments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalDatesBirthJ1Segments gets a reference to the given int32 and assigns it to the TotalDatesBirthJ1Segments field.
+func (o *TrailerRecord) SetTotalDatesBirthJ1Segments(v int32) {
+	o.TotalDatesBirthJ1Segments = &v
+}
+
+// GetTotalDatesBirthJ2Segments returns the TotalDatesBirthJ2Segments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalDatesBirthJ2Segments() int32 {
+	if o == nil || o.TotalDatesBirthJ2Segments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalDatesBirthJ2Segments
+}
+
+// GetTotalDatesBirthJ2SegmentsOk returns a tuple with the TotalDatesBirthJ2Segments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalDatesBirthJ2SegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalDatesBirthJ2Segments == nil {
+		return nil, false
+	}
+	return o.TotalDatesBirthJ2Segments, true
+}
+
+// HasTotalDatesBirthJ2Segments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalDatesBirthJ2Segments() bool {
+	if o != nil && o.TotalDatesBirthJ2Segments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalDatesBirthJ2Segments gets a reference to the given int32 and assigns it to the TotalDatesBirthJ2Segments field.
+func (o *TrailerRecord) SetTotalDatesBirthJ2Segments(v int32) {
+	o.TotalDatesBirthJ2Segments = &v
+}
+
+// GetTotalTelephoneNumbersAllSegments returns the TotalTelephoneNumbersAllSegments field value if set, zero value otherwise.
+func (o *TrailerRecord) GetTotalTelephoneNumbersAllSegments() int32 {
+	if o == nil || o.TotalTelephoneNumbersAllSegments == nil {
+		var ret int32
+		return ret
+	}
+	return *o.TotalTelephoneNumbersAllSegments
+}
+
+// GetTotalTelephoneNumbersAllSegmentsOk returns a tuple with the TotalTelephoneNumbersAllSegments field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *TrailerRecord) GetTotalTelephoneNumbersAllSegmentsOk() (*int32, bool) {
+	if o == nil || o.TotalTelephoneNumbersAllSegments == nil {
+		return nil, false
+	}
+	return o.TotalTelephoneNumbersAllSegments, true
+}
+
+// HasTotalTelephoneNumbersAllSegments returns a boolean if a field has been set.
+func (o *TrailerRecord) HasTotalTelephoneNumbersAllSegments() bool {
+	if o != nil && o.TotalTelephoneNumbersAllSegments != nil {
+		return true
+	}
+
+	return false
+}
+
+// SetTotalTelephoneNumbersAllSegments gets a reference to the given int32 and assigns it to the TotalTelephoneNumbersAllSegments field.
+func (o *TrailerRecord) SetTotalTelephoneNumbersAllSegments(v int32) {
+	o.TotalTelephoneNumbersAllSegments = &v
+}
+
+func (o TrailerRecord) MarshalJSON() ([]byte, error) {
+	toSerialize := map[string]interface{}{}
+	if o.BlockDescriptorWord != nil {
+		toSerialize["blockDescriptorWord"] = o.BlockDescriptorWord
+	}
+	if true {
+		toSerialize["recordDescriptorWord"] = o.RecordDescriptorWord
+	}
+	if true {
+		toSerialize["recordIdentifier"] = o.RecordIdentifier
+	}
+	if o.TotalBaseRecords != nil {
+		toSerialize["totalBaseRecords"] = o.TotalBaseRecords
+	}
+	if o.TotalStatusCodeDF != nil {
+		toSerialize["totalStatusCodeDF"] = o.TotalStatusCodeDF
+	}
+	if o.TotalConsumerSegmentsJ1 != nil {
+		toSerialize["totalConsumerSegmentsJ1"] = o.TotalConsumerSegmentsJ1
+	}
+	if o.TotalConsumerSegmentsJ2 != nil {
+		toSerialize["totalConsumerSegmentsJ2"] = o.TotalConsumerSegmentsJ2
+	}
+	if o.BlockCount != nil {
+		toSerialize["blockCount"] = o.BlockCount
+	}
+	if o.TotalStatusCodeDA != nil {
+		toSerialize["totalStatusCodeDA"] = o.TotalStatusCodeDA
+	}
+	if o.TotalStatusCode05 != nil {
+		toSerialize["totalStatusCode05"] = o.TotalStatusCode05
+	}
+	if o.TotalStatusCode11 != nil {
+		toSerialize["totalStatusCode11"] = o.TotalStatusCode11
+	}
+	if o.TotalStatusCode13 != nil {
+		toSerialize["totalStatusCode13"] = o.TotalStatusCode13
+	}
+	if o.TotalStatusCode61 != nil {
+		toSerialize["totalStatusCode61"] = o.TotalStatusCode61
+	}
+	if o.TotalStatusCode62 != nil {
+		toSerialize["totalStatusCode62"] = o.TotalStatusCode62
+	}
+	if o.TotalStatusCode63 != nil {
+		toSerialize["totalStatusCode63"] = o.TotalStatusCode63
+	}
+	if o.TotalStatusCode64 != nil {
+		toSerialize["totalStatusCode64"] = o.TotalStatusCode64
+	}
+	if o.TotalStatusCode65 != nil {
+		toSerialize["totalStatusCode65"] = o.TotalStatusCode65
+	}
+	if o.TotalStatusCode71 != nil {
+		toSerialize["totalStatusCode71"] = o.TotalStatusCode71
+	}
+	if o.TotalStatusCode78 != nil {
+		toSerialize["totalStatusCode78"] = o.TotalStatusCode78
+	}
+	if o.TotalStatusCode80 != nil {
+		toSerialize["totalStatusCode80"] = o.TotalStatusCode80
+	}
+	if o.TotalStatusCode82 != nil {
+		toSerialize["totalStatusCode82"] = o.TotalStatusCode82
+	}
+	if o.TotalStatusCode83 != nil {
+		toSerialize["totalStatusCode83"] = o.TotalStatusCode83
+	}
+	if o.TotalStatusCode84 != nil {
+		toSerialize["totalStatusCode84"] = o.TotalStatusCode84
+	}
+	if o.TotalStatusCode88 != nil {
+		toSerialize["totalStatusCode88"] = o.TotalStatusCode88
+	}
+	if o.TotalStatusCode89 != nil {
+		toSerialize["totalStatusCode89"] = o.TotalStatusCode89
+	}
+	if o.TotalStatusCode93 != nil {
+		toSerialize["totalStatusCode93"] = o.TotalStatusCode93
+	}
+	if o.TotalStatusCode94 != nil {
+		toSerialize["totalStatusCode94"] = o.TotalStatusCode94
+	}
+	if o.TotalStatusCode95 != nil {
+		toSerialize["totalStatusCode95"] = o.TotalStatusCode95
+	}
+	if o.TotalStatusCode96 != nil {
+		toSerialize["totalStatusCode96"] = o.TotalStatusCode96
+	}
+	if o.TotalStatusCode97 != nil {
+		toSerialize["totalStatusCode97"] = o.TotalStatusCode97
+	}
+	if o.TotalECOACodeZ != nil {
+		toSerialize["totalECOACodeZ"] = o.TotalECOACodeZ
+	}
+	if o.TotalEmploymentSegments != nil {
+		toSerialize["totalEmploymentSegments"] = o.TotalEmploymentSegments
+	}
+	if o.TotalOriginalCreditorSegments != nil {
+		toSerialize["totalOriginalCreditorSegments"] = o.TotalOriginalCreditorSegments
+	}
+	if o.TotalPurchasedToSegments != nil {
+		toSerialize["totalPurchasedToSegments"] = o.TotalPurchasedToSegments
+	}
+	if o.TotalMortgageInformationSegments != nil {
+		toSerialize["totalMortgageInformationSegments"] = o.TotalMortgageInformationSegments
+	}
+	if o.TotalPaymentInformationSegments != nil {
+		toSerialize["totalPaymentInformationSegments"] = o.TotalPaymentInformationSegments
+	}
+	if o.TotalChangeSegments != nil {
+		toSerialize["totalChangeSegments"] = o.TotalChangeSegments
+	}
+	if o.TotalSocialNumbersAllSegments != nil {
+		toSerialize["totalSocialNumbersAllSegments"] = o.TotalSocialNumbersAllSegments
+	}
+	if o.TotalSocialNumbersBaseSegments != nil {
+		toSerialize["totalSocialNumbersBaseSegments"] = o.TotalSocialNumbersBaseSegments
+	}
+	if o.TotalSocialNumbersJ1Segments != nil {
+		toSerialize["totalSocialNumbersJ1Segments"] = o.TotalSocialNumbersJ1Segments
+	}
+	if o.TotalSocialNumbersJ2Segments != nil {
+		toSerialize["totalSocialNumbersJ2Segments"] = o.TotalSocialNumbersJ2Segments
+	}
+	if o.TotalDatesBirthAllSegments != nil {
+		toSerialize["totalDatesBirthAllSegments"] = o.TotalDatesBirthAllSegments
+	}
+	if o.TotalDatesBirthBaseSegments != nil {
+		toSerialize["totalDatesBirthBaseSegments"] = o.TotalDatesBirthBaseSegments
+	}
+	if o.TotalDatesBirthJ1Segments != nil {
+		toSerialize["totalDatesBirthJ1Segments"] = o.TotalDatesBirthJ1Segments
+	}
+	if o.TotalDatesBirthJ2Segments != nil {
+		toSerialize["totalDatesBirthJ2Segments"] = o.TotalDatesBirthJ2Segments
+	}
+	if o.TotalTelephoneNumbersAllSegments != nil {
+		toSerialize["totalTelephoneNumbersAllSegments"] = o.TotalTelephoneNumbersAllSegments
+	}
+	return json.Marshal(toSerialize)
+}
+
+type NullableTrailerRecord struct {
+	value *TrailerRecord
+	isSet bool
+}
+
+func (v NullableTrailerRecord) Get() *TrailerRecord {
+	return v.value
+}
+
+func (v *NullableTrailerRecord) Set(val *TrailerRecord) {
+	v.value = val
+	v.isSet = true
+}
+
+func (v NullableTrailerRecord) IsSet() bool {
+	return v.isSet
+}
+
+func (v *NullableTrailerRecord) Unset() {
+	v.value = nil
+	v.isSet = false
+}
+
+func NewNullableTrailerRecord(val *TrailerRecord) *NullableTrailerRecord {
+	return &NullableTrailerRecord{value: val, isSet: true}
+}
+
+func (v NullableTrailerRecord) MarshalJSON() ([]byte, error) {
+	return json.Marshal(v.value)
+}
+
+func (v *NullableTrailerRecord) UnmarshalJSON(src []byte) error {
+	v.isSet = true
+	return json.Unmarshal(src, &v.value)
+}
+
+
