@@ -7,8 +7,9 @@ package lib
 // General segment interface
 type Segment interface {
 	Name() string
-	Parse(string) (int, error)
+	Parse([]byte) (int, error)
 	String() string
+	Bytes() []byte
 	Validate() error
 	Length() int
 }
