@@ -4,4 +4,4 @@
 
 package metro2
 
-const Version = "v0.3.3"
+const Version = "v0.5.0"
